% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_assign_stars.R
\name{assign_stars}
\alias{assign_stars}
\title{Assign Significance Stars to P-values}
\usage{
assign_stars(p_value)
}
\arguments{
\item{p_value}{A numeric value representing the p-value for which significance stars are to be assigned.}
}
\value{
A character string containing the significance stars ('***', '**', '*', or '').
}
\description{
This function takes a p-value as an argument and returns a string of asterisks
indicating the significance level. The asterisks follow the common conventions:
'***' for p <= 0.001, '**' for p <= 0.01, '*' for p <= 0.05, and '' for non-significant p-values.
}
\examples{
assign_stars(0.0002) # returns '***'
assign_stars(0.01)   # returns '**'
assign_stars(0.05)   # returns '*'
assign_stars(0.2)    # returns ''

}
