% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_density.R
\name{avg_density}
\alias{avg_density}
\title{Calculates Average Fish Density and bootstrap confidence intervals}
\usage{
avg_density(counts, area, iter = NULL, conf = NULL)
}
\arguments{
\item{counts}{dataframe or tibble of counts (can also be a vector if one species)}

\item{area}{vector of areas corresponding to observations (rows) in counts}

\item{iter}{optional number of iterations for creating confidence intervals}

\item{conf}{optional confidence level (0-1) (If null will return average density for each bootstrap sample)}
}
\description{
Calculates Fish Densities and optionally bootstrap confidence intervals.
Returns a dataframe or tibble of densities for each species (depending on class input as counts)
If iter and conf are not NULL than row order will be lower bound, calculated average, upper bound
If iter isn not NULL but conf is NULL, than the mean from each bootstrap sample will be returned
}
