% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cvision.R
\name{bin_cvision}
\alias{bin_cvision}
\title{Collapses merged C-vision csv to time bins}
\usage{
bin_cvision(merged_cvision_csv, start_vid, end_vid, bin_size = 15,
  vid_length = 1)
}
\arguments{
\item{merged_cvision_csv}{output of merge_cvision_csv (as a dataframe or tibble)}

\item{start_vid}{video to start output table with}

\item{end_vid}{video to end output table with}

\item{bin_size}{length of bins in seconds}

\item{vid_length}{length of video in minutes}
}
\description{
Collapses merged C-vision csv to time bins
}
