% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{binning}
\title{binning for matrix-like object}
\usage{
binning(x, size = 2L, ...)
}
\arguments{
\item{x}{A matrix-like object}

\item{size}{Binning size (default: `2L`)}

\item{...}{Discarded}
}
\description{
binning for matrix-like object
}
