% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPI.R
\name{BPI}
\alias{BPI}
\title{Calculates Bathymetric Position Index}
\usage{
BPI(bathy, inner_rad, outer_rad, rad_units = "cell", na.rm = FALSE,
  pad = FALSE)
}
\arguments{
\item{bathy}{bathymetry raster}

\item{inner_rad}{Inner radius of annulus}

\item{outer_rad}{Outer radius of annulus}

\item{pad}{logical. If TRUE, additional 'virtual' rows and columns are padded to x such that there are no edge effects. Only applicable if na.rm=TRUE. Value of virtual cells is set to NA. Default is FALSE}
}
\description{
Calculates Bathymetric Position Index (the difference between a central pixel and the mean of the values in the specified annulus). Positive values inidcate a crest, while negative values indicate a depression. 0 value is flat.
CURRENTLY DOES NOT MATCH VALUES FROM BENTHIC TERRAIN MODELLER
}
