% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broken_stick.R
\name{broken_stick}
\alias{broken_stick}
\title{Returns Eigenvalues expected by random data}
\usage{
broken_stick(n_variables)
}
\arguments{
\item{n_variables}{number of variables}
}
\description{
Returns Eigenvalues expected by random data for each component as determined by a broken-stick distribution. Useful for determining the number of interpretable Principal Components. When actual eigenvalues fall below this curve they are no longer considered interpretable.
}
