% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bearing.R
\name{calc_bearing}
\alias{calc_bearing}
\title{Calculates bearing between points}
\usage{
calc_bearing(p1, p2)
}
\arguments{
\item{p1}{Start point. Specified as c(Easting, Northing)}

\item{p2}{End Point. Specified as c(Easting, Northing)}
}
\description{
Calculates bearing between two points in UTM coordinates
}
