% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_density.R
\name{calc_dens}
\alias{calc_dens}
\title{Helper function to calculate fish densities}
\usage{
calc_dens(counts, area)
}
\arguments{
\item{counts}{dataframe of counts}

\item{area}{vector of areas corresponding to observations (rows) in counts}
}
\description{
calculate average fish densities
}
