% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{calc_dist}
\alias{calc_dist}
\title{Calculates distance CBASS has traveled}
\usage{
calc_dist(speed, start_time = NULL, end_time = NULL,
  speed_units = "knots", duration = NULL)
}
\arguments{
\item{speed}{speed of CBASS}

\item{start_time}{start time}

\item{end_time}{end time}

\item{speed_units}{option of "knots" (default) or "m/s"}

\item{duration}{instead of start/end time can also supply the duration in seconds}
}
\description{
Calculates distance (in meters) CBASS has traveled over a given duration based on speed and time
}
