% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_entropy.R
\name{calc_entropy}
\alias{calc_entropy}
\title{Calculates Shannon entropy of a raster stack of class probabilities.}
\usage{
calc_entropy(raster_stack, base = exp(1))
}
\arguments{
\item{raster_stack}{A raster stack/brick object containing layers that correspond to class probabilities}

\item{base}{base of logarithm (default is natural logarithm)}
}
\description{
Calculates Shannon entropy of a raster stack of class probabilities. Not to be confused with that requires categorical/integer raster values fo rlooking at agreement among an ensemble of models. Defaults to using the natural logarithm.
}
