% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_FOV.R
\name{calc_FOV}
\alias{calc_FOV}
\title{Calculates Field of View}
\usage{
calc_FOV(d = NULL, f = NULL, medium, type, pix = NULL,
  aspect_ratio = NULL, pixel_size, n_air = 1.000277, n_sea = 4/3)
}
\arguments{
\item{d}{length of diagonal of camera sensor (must be same units as f)}

\item{f}{focal length of lens (must be same units as d or pixel size)}

\item{medium}{"air" or "seawater"}

\item{type}{Type of field of view. Either "horiz", "vert", or "diag" (diag does not need pix or aspect ratio)}

\item{pix}{width and height of frame in number of pixels c(width, height)}

\item{aspect_ratio}{aspect ratio e.g. c(16,9)}

\item{pixel_size}{pixel size of camera sensor (must be same units as f)}

\item{n_air}{index of refraction for air (default = 1.000277)}

\item{n_sea}{index of refraction for seawater (default = 4/3)}
}
\description{
Calculates field of view (in degrees) based on camera and lens properties. Can calculate horizontal, vertical, and diagonal field of view in either air or seawater. If d is unknown, it can be approximated by pix and pixel_size. If d is known, either aspect_ratio or pix can be used.
}
