% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_layback.R
\name{calc_layback}
\alias{calc_layback}
\title{Calculates layback of CBASS}
\usage{
calc_layback(payout, depth, GPS_Source = "None", zeroed, cat_fact = 1)
}
\arguments{
\item{payout}{line out from boat (meters)}

\item{depth}{depth of CBASS (meters)}

\item{GPS_Source}{Adds offset forward/aft offset between trawl block and GPS source. Must be "None" or "GNSS_Fwd" or "CoG" (Center of Gravity)}

\item{zeroed}{"water" or "block"}

\item{cat_fact}{catenary factor (from 0-1)}
}
\description{
Calculates layback of CBASS in meters
}
