% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_breaks.R
\name{calculate_breaks}
\alias{calculate_breaks}
\title{calculate_breaks: Calculate Axis Breaks}
\usage{
calculate_breaks(limits)
}
\arguments{
\item{limits}{Numeric vector of length 2, specifying the lower and upper limits of the axis.}
}
\value{
Numeric vector containing the calculated breaks within the specified limits.
}
\description{
Internal function that calculates axis breaks by determining an appropriate interval ('by') value and applying it within the specified limits. It leverages the `get_adaptive_by` function to adjust the 'by' value based on the range of the data.
}
\details{
This function calculates the breaks for an axis based on the limits provided. It uses an adaptive approach to determine suitable intervals between breaks, aiming for a visually appealing and informative axis scale.
}
