% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_minor_breaks.R
\name{calculate_minor_breaks}
\alias{calculate_minor_breaks}
\title{#' calculate_minor_breaks: Calculate Minor Axis Breaks}
\usage{
calculate_minor_breaks(limits)
}
\arguments{
\item{limits}{Numeric vector of length 2, specifying the lower and upper limits of the axis.}
}
\value{
Numeric vector containing the calculated minor breaks, or NULL if the addition of minor breaks would result in an overcrowded axis.
}
\description{
Internal function that calculates minor axis breaks. It uses the main axis breaks (calculated using `calculate_breaks`) as a basis and introduces additional minor breaks between them. The function aims to improve plot readability without overcrowding the axis. The number of minor breaks is dynamically adjusted based on the spacing of the major breaks.
}
\details{
This function calculates minor axis breaks based on the main breaks to enhance the readability of plots. The minor breaks are calculated to fall between the major breaks, with the number of minor breaks adjusted to avoid overcrowding the axis.
}
