% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_camera_table.R
\name{check_camera_table}
\alias{check_camera_table}
\title{QC for camera table}
\usage{
check_camera_table(camera_table, display_warning = TRUE)
}
\arguments{
\item{camera_table}{camera table}

\item{display_warning}{Logical indicating whether a warning should be displayed}
}
\description{
Sequentially checks if camera table has one pattern, if it has been reset, and if it is jumbled (arranging by time and arranging by frame number are different). Returns what the first issue encounted is, which is either "pattern", "reset," or "jumbled," or "none.
}
