% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cvision.R
\name{check_cvision}
\alias{check_cvision}
\title{Check if any C-Vision csv's are missing}
\usage{
check_cvision(expected, file_list = list.files(pattern = ".csv$"))
}
\arguments{
\item{expected}{vector of expected video numbers}

\item{file_list}{list of cvision csv files. If left blank will use all csvs in working directory}
}
\description{
Check if any C-Vision csv's are missing
}
