% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_jumbled.R
\name{check_jumbled}
\alias{check_jumbled}
\title{Check if camera table is jumbled}
\usage{
check_jumbled(camera_table, display_warning = TRUE)
}
\arguments{
\item{camera_table}{camera table}

\item{display_warning}{Logical indicating whether a warning should be displayed}
}
\description{
Check if camera table is jumbled (arranging by time and arranging by frame number are different). If true that may mean that the clock was resynched.
}
