% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pattern.R
\name{check_pattern}
\alias{check_pattern}
\title{Checks if camera table has more than one pattern}
\usage{
check_pattern(cam_filepath, display_warning = TRUE)
}
\arguments{
\item{cam_filepath}{vector of filepaths from blackfly camera}

\item{display_warning}{Logical indicating whether a warning should be displayed if issue detected}
}
\description{
Checks if camera table has more than one pattern (e.g. camera or image file extension type). TRUE indicates one pattern found which is what it should be.
}
