% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.r
\name{checkme}
\alias{checkme}
\title{Perform standard model fit checks}
\usage{
checkme(model, resp)
}
\arguments{
\item{model}{The object containing the fitted model.}

\item{resp}{The response variable in the fitted model.}
}
\description{
For a fitted model (lm etc.) performs standard model checking 
plots (residuals vs. fits, qq plot) etc.
}

