% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.randomised.r
\name{chisq.randomised}
\alias{chisq.randomised}
\title{Chi-squared contingency table by randomisation}
\usage{
chisq.randomised(observed, rands)
}
\arguments{
\item{observed}{A dataframe with a single column of observed frequencies}

\item{rands}{The number of randomisations to perform}
}
\description{
function to determine significance of chi-squared goodness 
of fit by randomisation to deal with low frequencies.
}

