% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_Cibersort.R
\name{CIBERSORT_deconv}
\alias{CIBERSORT_deconv}
\title{CIBERSORT_deconv}
\usage{
CIBERSORT_deconv(mixture_file, 
sig_matrix = "LM22", 
perm=1, 
QN=FALSE)
}
\arguments{
\item{mixture_file}{Heterogenous mixed expression.}

\item{sig_matrix}{Gene expression matrix from isolated cells.}

\item{perm}{Number of permutations. Default is 1.}

\item{QN}{Perform quantile normalization or not (TRUE/FALSE). Recommend FALSE for RNA-seq data, TRUE for microarray data.}
}
\value{
A data.frame containing predicted immune cell proportions.
}
\description{
A function to deconvolute bulk expression
}
\author{
GM.W
}
