% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_window.R
\name{circle_window}
\alias{circle_window}
\title{Creates circular focal window}
\usage{
circle_window(radius, resolution = NA, unit = "cell", type = 1)
}
\arguments{
\item{radius}{radius of circular window}

\item{resolution}{resolution of intended raster layer (one number or a vector of length 2). Only necessary if unit= "map"}

\item{unit}{unit for radius. Either "cell" (number of cells, the default) or "map" for map units (e.g. meters).}

\item{type}{Either 1 or 2. If type = "1" then weights will be 0 or 1. if type ="2", weights will sum to 1. Default is 1.}
}
\description{
Creates circular focal window around central pixel.
https://desktop.arcgis.com/en/arcmap/10.3/analyze/arcpy-spatial-analyst/nbrcircle-class.htm
}
