% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_col_name.R
\name{clean_col_name}
\alias{clean_col_name}
\title{Clean Column Names}
\usage{
clean_col_name(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
A data frame
}
\description{
Clean Column Names
}
\examples{
df <- data.frame(tmPP = c(1:5), YnUU = c(1:5))
df <- clean_col_name(df)
}
