% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clna.R
\name{clna}
\alias{clna}
\title{Summary statistics for missing values and classes of a dataframe}
\usage{
clna(data, all_vars = F, vars = names(data))
}
\arguments{
\item{data}{The dataframe to be analyzed.}

\item{all_vars}{Logical, if FALSE only variables with NAs are returned, if TRUE all variables are returned. Default is FALSE.}

\item{vars}{Character vector of variable names to be analyzed. Default is all variables in the dataframe.}
}
\value{
A dataframe containing the following columns:
\itemize{
  \item Variable: The name of the variable.
  \item class: The class type of the variable.
  \item n_NAs: The number of missing values in the variable.
  \item perc_NAs: The percentage of missing values in the variable.
}
}
\description{
This function returns a summary dataframe that includes information about the number
and percentage of missing values (NAs) as well as the classes of variables in the dataframe.
}
\examples{
df <- data.frame(a = c(1, 2, 3, NA), b = c("a", "b", "c", "d"), c = c(TRUE, TRUE, FALSE, NA))
clna(df)
clna(df, all_vars = TRUE)
clna(df, vars = c("a", "c"))

}
