% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_bins.R
\name{collapse_bins}
\alias{collapse_bins}
\title{Collapse Percent Cover Bins to a new specified range}
\usage{
collapse_bins(percent_cover, new_bins)
}
\arguments{
\item{percent_cover}{a character vector of percent cover data in ranges formatted as "lower-upper" or a single number a a character}

\item{new_bins}{vector of new bins formatted as a string with (eg c("0", "1-5", "6-100")). Ranges must be in the format "lower-upper" or a single number as a character.}
}
\description{
Takes a vector of percent cover data and a vector of new bin ranges to collapse the data
}
