% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_rasters.R
\name{combine_rasters}
\alias{combine_rasters}
\title{Combines multiple rasters into one}
\usage{
combine_rasters(raster_list, ref_rast = NULL, desired_origin,
  desired_res, desired_crs, method = "bilinear",
  display_progress = TRUE, normalize = FALSE)
}
\arguments{
\item{raster_list}{a list of rasters}

\item{ref_rast}{reference raster with desired properties (Alternatively can supply desired_origin, desired_res, and desired_crs)}

\item{desired_origin}{desired origin of output raster as a vector with length 2 (x,y)}

\item{desired_res}{desired resolution of output raster. Either an integer or a vector of length 2 (x,y)}

\item{desired_crs}{desired coordinate reference system of output raster (CRS class)}

\item{method}{resampling method. Either "bilinear" for bilinear interpolation (the default), or "ngb" for using the nearest neighbor}

\item{display_progress}{Logical specifying whether or not to indicate progress}

\item{normalize}{Logical (False by Default) indicating whether to z-score normalize raster layer before merge (Useful for merging backscatter)}

\item{rast}{raster to be reprojected or resampled}
}
\description{
Combines multiple rasters into one by using merge after first reprojecting or resampling and aligning rasters by matching them up with a a specified origin, resolution, and coordinate reference system, or that of a reference raster.
}
