% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_cond_entropy.R
\name{cond_entropy}
\alias{cond_entropy}
\title{Calculate Conditional Entropy of Features Relative to a Target Feature}
\usage{
cond_entropy(dat, target, exclude = NULL)
}
\arguments{
\item{dat}{A data frame containing the dataset.}

\item{target}{The name of the target feature.}

\item{exclude}{A vector of feature names to exclude from the analysis.}
}
\value{
A sorted data frame with features and their conditional entropy.
}
\description{
This function computes the conditional entropy for each feature in a dataset,
relative to a specified target feature.
}
\details{
The source code is not that complicated. So if you are not sure if this function fits your needs, run 'cons_entropy' without brackets in your console and read the code by yourself.
}
\examples{
# Add example usage here

}
