% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_CBASS_db.R
\name{connect_CBASS_db}
\alias{connect_CBASS_db}
\title{Connect to a CBASS database}
\usage{
connect_CBASS_db(dbname, user = "dataReader", password = "12345")
}
\arguments{
\item{dbname}{character string with name of database. Names of databases can be accessed with getCBASSdbNames(). Database names are "information_schema", "cLayback", "cbass1", "cbass2_nuc1", "cbass2_nuc2", "cbass2_nuc3", "mysql", "performance_schema", "sys".}

\item{user}{username}

\item{password}{password}
}
\description{
Connect to a CBASS database. Names of databases can be accessed with getCBASSdbNames(). Must be connected through sars-tunnel.
}
