% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.r
\name{consum}
\alias{consum}
\title{Sets contrasts to 'sum' option}
\usage{
consum()
}
\description{
For linear models, sets the contrasts option from the 
default to 'sum' i.e. sum to zero contrasts. This gives output consistent 
with Minitab etc.
}

