% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.r
\name{contre}
\alias{contre}
\title{Sets contrasts to 'treatment' option}
\usage{
contre()
}
\description{
For linear models, sets the contrasts option to 'treatment' 
i.e. contrasts are set to compare each level with the baseline level 
(non-orthogonal).
}

