% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_directories.R
\name{create_directories}
\alias{create_directories}
\title{Create Directories in the Working Directory}
\usage{
create_directories(dir_names = c("data", "functions", "plots", "results"))
}
\arguments{
\item{dir_names}{A character vector specifying the names of the directories to be created. Default is c('data', 'plots', 'results').}
}
\value{
Invisible NULL. Directories are created for side-effects.
}
\description{
This function creates specified directories in the current working directory.
By default, it creates 'data', 'plots', and 'results' folders.
}
\examples{
\dontrun{
# Create default directories: 'data', 'plots', 'results'
create_directories()

# Create custom directories: 'custom1', 'custom2'
create_directories(c('custom1', 'custom2'))
}
}
