% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cscore.R
\name{cscore}
\alias{cscore}
\title{Calculates normalized pairwise checkerboard C-scores}
\usage{
cscore(pres_abs)
}
\arguments{
\item{pres_abs}{binary dataframe or tibble of presence/absence data where species are columns and observations are rows. Presence is indicated by 1 and absence is indicated by 0.}
}
\description{
Calculates normalized checkerboard C-scores based on Stone and Roberts 1990 and then standardizes them to a value between 0 and 1 by dividing by the maxiumum possible C-score for that pair of species (Groundfish Management Team 2013 http://www.pcouncil.org/wp-content/uploads/F8b_GMT_JUN2013BB.pdf)
Outputs a square symmetric species by species dataframe containing pairwise c-scores.
0 indicates complete overlap. 1 indicates complete segregation.
> .7 indicates strong segregation and <.3 indicates strong overlap (Groundfish Management Team 2013)
}
