% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dd}
\alias{dd}
\title{Double density plots}
\usage{
dd(x, g)
}
\arguments{
\item{x}{A numeric vector containing the values.}

\item{g}{A vector containing two unique values, indicating the groups.}
}
\value{
An object of class `ggplot`.
}
\description{
A double density plot shows the densities of two groups. Both have an AUC
of 1. It can be used to compare the results of a predictive model, but
also the predictive potential of a single feature.
}
\examples{
x <- c(rnorm(100,1), rnorm(100, 2))
g <- rep(letters[1:2], each = 100)
dd(x, g)
}

