% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_DEG.R
\name{DEG_calculate}
\alias{DEG_calculate}
\title{DEG_calculate}
\usage{
DEG_calculate(exprset = count,
 dtype = "count",
 method = "DESeq2",
 gene = "NEIL3")
}
\arguments{
\item{exprset}{The exprset to perform DEG analysis}

\item{dtype}{The type of exprset, either "count", "fpkm", "tpm", or "microarray"}

\item{method}{The method used to perform DEG analysis, either "DESeq2", "edgeR", or "limma"}

\item{gene}{The gene of interest. Param \code{gene} and param \verb{samples_*} should not given at the same time.}

\item{percent}{Float, (0,1]. The percent of samples used in this analysis. e.g. percent=0.8 equals to top 40\% and tail 40\% samples. Default is 1.}

\item{samples_con}{The vector of samples treated as control.}

\item{samples_case}{The vector of samples treated as case.}
}
\value{
The data.frame of DEGs
}
\description{
Calculate Differentially expressed genes between
high- and low- gene expressed samples or two specific groups.
}
\author{
GM.W
}
