% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_DEG.R
\name{DEG_volcano_plot}
\alias{DEG_volcano_plot}
\title{DEG_volcano_plot}
\usage{
DEG_volcano_plot(DEG_res,
 method="DESeq2",
 logFC_threshold=1, 
 pvalue_threshold=0.05, 
 label_num_per=10,
 title=NULL)
}
\arguments{
\item{DEG_res}{The result of \code{DEG_calculate}}

\item{method}{The method used to perform DEG analysis}

\item{logFC_threshold}{The threshold of log2FoldChange. Default is 1.}

\item{pvalue_threshold}{The threshold of pvalue. Default is 0.05.}

\item{label_num_per}{The number of gene to be shown in each side.}

\item{title}{The title of the plot.}
}
\value{
A ggplot object.
}
\description{
Plot a volcano graph based on DEG result using tools in c("DESeq2","edgeR","limma_voom","limma","wilcox")
}
\examples{
DEG_volcano_plot(DEG_res,method=c"DESeq2")

}
\author{
GM.W
}
