% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degrees.R
\name{degrees}
\alias{degrees}
\title{Converts a measurement in radians to degrees}
\usage{
degrees(rad)
}
\arguments{
\item{rad}{measurement in radians}
}
\description{
Converts a measurement in radians to degrees
}
