% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stats.R
\name{desc_table}
\alias{desc_table}
\title{一键生成描述统计表}
\usage{
desc_table(data, group_var = NULL, digits = 1)
}
\arguments{
\item{data}{数据框}

\item{group_var}{分组变量名（字符串），如 "treatment"}

\item{digits}{小数位数，默认 1}
}
\value{
一个 tibble，包含各变量的分组统计结果
}
\description{
对数据框中的数值变量计算均值(SD)，分类变量计算n(\%)，
按指定分组变量分组汇总。
}
\examples{
data(mtcars)
mtcars$am <- factor(mtcars$am, labels = c("Auto", "Manual"))
desc_table(mtcars, group_var = "am")

}
