% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_enrichment&GSEA.R
\name{.gseaNb}
\alias{.gseaNb}
\title{.gseaNb}
\usage{
.gseaNb(
  object = NULL,
  subPlot = 3,
  lineSize = 0.8,
  geneSetID = NULL,
  rmSegment = FALSE,
  termWidth = 40,
  segCol = "red",
  addGene = NULL,
  geneCol = NULL,
  arrowAngle = 20,
  arrowLength = 0.2,
  arrowEnd = "first",
  arrowType = "closed",
  curveCol = c("#76BA99", "#EB4747", "#996699"),
  htCol = c("#08519C", "#A50F15"),
  rankCol = c("#08519C", "white", "#A50F15"),
  rankSeq = 5000,
  htHeight = 0.3,
  force = 20,
  max.overlaps = 50,
  geneSize = 4,
  newGsea = FALSE,
  addPoint = TRUE,
  newCurveCol = c("#336699", "white", "#993399"),
  newHtCol = c("#336699", "white", "#993399"),
  rmHt = FALSE,
  addPval = FALSE,
  pvalX = 0.9,
  pvalY = 0.9,
  pvalSize = 4,
  pCol = "grey30",
  pHjust = 1,
  rmPrefix = TRUE,
  nesDigit = 2,
  pDigit = 2,
  markTopgene = FALSE,
  topGeneN = 5,
  kegg = FALSE,
  legend.position = "right",
  add.geneExpHt = FALSE,
  exp = NULL,
  scale.exp = TRUE,
  sample.order = NULL,
  exp.col = c("blue", "white", "red"),
  ht.legend = TRUE,
  ght.relHight = 0.4,
  ght.geneText.size = 6,
  ght.facet = FALSE,
  ght.facet.scale = "free",
  termID.order = NULL,
  rank.gene = NULL,
  rank.gene.nudgey = 2
)
}
\description{
Plot GSEAcurves
}
