% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicValuesCollector.R
\name{DynamicValuesCollector}
\alias{DynamicValuesCollector}
\title{DynamicValuesCollector Class}
\description{
A class for storing and retrieving various metrics for different models.
}
\examples{
\dontrun{
collector <- DynamicValuesCollector$new(c("R2", "Adj_R2", "RMSE"))
collector$add("model1", c(R2 = 0.9, Adj_R2 = 0.89, RMSE = 0.1))
collector$display_sorted("R2")
}
Constructor

Initializes an empty list to store models and sets the metric names.

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{models}}{A list to store models along with their metrics.}

\item{\code{metric_names}}{A vector of names for the metrics to be stored.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DynamicValuesCollector-new}{\code{DynamicValuesCollector$new()}}
\item \href{#method-DynamicValuesCollector-add}{\code{DynamicValuesCollector$add()}}
\item \href{#method-DynamicValuesCollector-remove}{\code{DynamicValuesCollector$remove()}}
\item \href{#method-DynamicValuesCollector-display_sorted}{\code{DynamicValuesCollector$display_sorted()}}
\item \href{#method-DynamicValuesCollector-clone}{\code{DynamicValuesCollector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynamicValuesCollector-new"></a>}}
\if{latex}{\out{\hypertarget{method-DynamicValuesCollector-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicValuesCollector$new(metric_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metric_names}}{A vector of names for the metrics to be stored.
Add a Model

Adds a model along with its metrics.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynamicValuesCollector-add"></a>}}
\if{latex}{\out{\hypertarget{method-DynamicValuesCollector-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicValuesCollector$add(model_name, metrics)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{The name of the model.}

\item{\code{metrics}}{A named vector or list containing the metrics.
Remove a Model

Removes a model based on its name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynamicValuesCollector-remove"></a>}}
\if{latex}{\out{\hypertarget{method-DynamicValuesCollector-remove}{}}}
\subsection{Method \code{remove()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicValuesCollector$remove(model_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{The name of the model to remove.
Display Sorted Models

Sorts and displays the models based on a specific metric.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynamicValuesCollector-display_sorted"></a>}}
\if{latex}{\out{\hypertarget{method-DynamicValuesCollector-display_sorted}{}}}
\subsection{Method \code{display_sorted()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicValuesCollector$display_sorted(sort_by, as_dataframe = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sort_by}}{The name of the metric to sort by.}

\item{\code{as_dataframe}}{Logical, whether to return the result as a data frame. Default is TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame or printed output based on the `as_dataframe` parameter.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynamicValuesCollector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DynamicValuesCollector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicValuesCollector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
