% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_enrichment&GSEA.R
\name{enrich_analysis}
\alias{enrich_analysis}
\title{enrich_analysis}
\usage{
enrich_analysis(DEG_res, 
n_top = 200,
method = "DESeq2",
signatures = c("GO","KEGG","HALLMARK","REACTOME","TF","miRNA"),
padj_threshold = 0.05)
}
\arguments{
\item{DEG_res}{The result of DEG_calculate}

\item{n_top}{The top n genes were selected for Hypergeometric distribution (GO & KEGG).}

\item{method}{The method of \code{DEG_calculate}. It should be one of c("DESeq2","edgeR","limma_voom","limma","wilcox").}

\item{signatures}{The signatures for GSEA. Default is all of c("GO","KEGG","HALLMARK","REACTOME","TF","miRNA").}

\item{padj_threshold}{The p-adj threshold of DEGs.}
}
\value{
A object of "enrich" containing the results of Hypergeometric distribution and GSEA, as well as parameters used.
}
\description{
A function to conduct enrichment analysis including Hypergeometric distribution and GSEA.
}
\author{
GM.W
}
