% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_info.R
\name{environment_info}
\alias{environment_info}
\title{A function to print our session info}
\usage{
environment_info(msg)
}
\arguments{
\item{msg}{The message to be printed}
}
\value{
print on screen
}
\description{
A function to print our session info
}
\examples{
environment_info("this is a message to be printed on the screen")
}
