% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_window.R
\name{extract_window}
\alias{extract_window}
\title{Extracts relevant window from raster based on position of central pixel and window size}
\usage{
extract_window(r, w = c(3, 3), idx)
}
\arguments{
\item{r}{a raster}

\item{w}{A vector of length 2 specifying the dimensions of the rectangular window to use where the first number is the number of rows and the second number is the number of columns. Window size must be an odd number. If a single value is supplied it will be used for both the number of rows and columns.}

\item{idx}{A vector of length 2 representing the row and column position of the central pixel of the window you wish to extract from raster r}
}
\value{
A matrix containg the extracted window around the central pixel from raster r
}
\description{
Extracts relevant window from raster based on position of central pixel and window size
}
