% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastOddsRatio.R
\name{FastOddsRatio}
\alias{FastOddsRatio}
\title{FastOddsRatio}
\usage{
FastOddsRatio(x, y)
}
\arguments{
\item{x}{A numeric probability.}

\item{y}{A numeric probability.}
}
\value{
the odds ratio.
}
\description{
Calculate odds ration from two probability x and y doing the ration of odds(x) on odds(y)
}
\examples{
FastOddsRatio(0.3, 0.1)
}
