% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_cell_cell_communications.R
\name{filter_lr_paths}
\alias{filter_lr_paths}
\title{filter_lr_paths}
\usage{
filter_lr_paths(
  st_dec_data,
  species = "Human",
  lrpairs = NULL,
  pathways = NULL,
  max_hop = NULL,
  if_doParallel = T,
  use_n_cores = NULL
)
}
\arguments{
\item{st_dec_data}{Deconvoluted Spatial matrix.}

\item{species}{Character. "Human" or "Mouse". Default is "Human".}

\item{lrpairs}{Ligand-Receptor interaction pairs. Default is \code{SpaTalk::lrpairs}.}

\item{pathways}{Pathways with source and destination. Default is \code{SpaTalk::pathways}.}

\item{max_hop}{Numeric. The maximum hop to calculate neighboring cell-cell interaction. Default is 4.}

\item{if_doParallel}{Logical. Whether to use multi-cores to accelerate the process. Default is True.}

\item{use_n_cores}{Numeric. How many cores to use for this process. Default is (all cores - 2).}
}
\value{
A list consisting of (lrpairs, pathways, max_hop)
}
\description{
find out the LR pairs with the receptors have
at least 1 downstream TF and pathways.
}
\examples{
lr_paths_filtered <- filter_lr_paths(st_dec_data)
lrpairs_filtered <- lr_paths_filtered$lrpairs
pathways_filtered <- lr_paths_filtered$pathways
max_hop_used <- lr_paths_filtered$max_hop

}
\author{
GM.W
}
