% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findkappa.R
\name{find.kappa}
\alias{find.kappa}
\title{function for find coefficient of tail of GARCH process}
\usage{
find.kappa(l1, l2, precision = 0.01, inov = "norm", df = 1)
}
\arguments{
\item{l1}{lambda_1, coefficient of ARCH term}

\item{l2}{lambda_2, coefficient of GARCH term}

\item{precision}{precision level of result}

\item{inov}{innovation distribution of GARCH process}

\item{df}{degree of freedom if the innovation is Student's t}
}
\value{
numerical
}
\description{
function for find coefficient of tail of GARCH process
}
\examples{
#' find.kappa(0.1,0.85,precision = 0.001,inov = "norm"),
#' find.kappa(0.1,0.85,precision = 0.001,inov = "std",df = 5)
}

