% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_cell_cell_communications.R
\name{find_CCI}
\alias{find_CCI}
\title{find_CCI}
\usage{
find_CCI(st_dec_data, st_dec_meta,
  lr_paths_filtered,
  celltype_sender, celltype_receiver,
  n_neighbor = 10, min_pairs = 5, min_pairs_ratio = 0,
  per_num = 1000, pvalue = 0.05, co_exp_ratio = 0.1,
  if_doParallel = TRUE, use_n_cores = NULL)
}
\arguments{
\item{st_dec_data}{Deconvoluted spatial matrix.}

\item{st_dec_meta}{The metadata whose row names are cells and contains cell, x, y, and celltype.}

\item{lr_paths_filtered}{The result of \code{filter_lr_paths}, which contains filtered lrpairs, pathways, and max_hop.}

\item{celltype_sender}{Character. The sender cell type.}

\item{celltype_receiver}{Character. The receiver cell type.}

\item{n_neighbor}{Numeric. Number of neighbor cells to select as the proximal cell-cell pair. Default is 10.}

\item{min_pairs}{Numeric. Min proximal cell-cell pairs between for sending and receiving cell types. Default is 5.}

\item{min_pairs_ratio}{Min proximal cell-cell pairs ratio between for sending and receiving cell types. Default is 0.}

\item{per_num}{Number of repeat times for permutation test. Default is 1000.}

\item{pvalue}{Include the significantly proximal LR pairs with this cutoff of p value from permutation test. Default is 0.05.}

\item{co_exp_ratio}{Min cell ratio in receiving cells with co-expressed source and target genes for predicting the downstream pathway activity.}

\item{if_doParallel}{Use doParallel. Default is TRUE.}

\item{use_n_cores}{Number of CPU cores to use. Default is all cores - 2.}
}
\value{
A list consisting of (lrpair, tf, cell_pair, sender_receiver, params)
}
\description{
Find out the

找到receptor中共表达的TF和target

找到所有的TF及其所在的hop层数

找出每个TF下游激活的target数目

random walk检验，计算TF score.

Calculate the distance between cells

Find cell-cell pairs in neighborhood.

Find coexpressed Ligand-Receptor pairs in sender-receiver cells and calculate P-value.

Find downstream activated TF and targets of receptors.

Calculate the communication score between sender and receiver.
}
\examples{
CCI_res <- find_CCI(st_dec_data = SpaTalk_dec_sc,st_dec_meta = SpaTalk_cellmeta,
  lr_paths_filtered = lr_paths_filtered,
  celltype_sender = "Fibroblasts", celltype_receiver = "Cancer_clone_A",
  n_neighbor = 10)
  lrpair <- CCI_res$lrpair
  TF <- CCI_res$tf
  cell_pair <- CCI_res$cell_pair

  @author GM.W

}
