% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_fpkm2tpm.R
\name{fpkm2tpm}
\alias{fpkm2tpm}
\title{fpkm2tpm}
\usage{
as.data.frame(apply(fpkm, 2, fpkm2tpm))
}
\arguments{
\item{fpkm}{The expression matrix in the format of \code{FPKM}.}
}
\value{
A matrix of TPM.
}
\description{
To convert a fpkm matrix to a tpm matrix
}
\examples{
as.data.frame(apply(fpkm, 2, fpkm2tpm))

}
\author{
Not Me.
}
