% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread2.R
\name{fread2}
\alias{fread2}
\title{transposing fread}
\usage{
fread2(input, transposing = TRUE, remove_null = TRUE, ss = c("excel",
  "libre"), args1 = list(), args2 = list())
}
\arguments{
\item{input}{input data}

\item{transposing}{transpose data frame when rows indicate variables and cols indicate observations}

\item{remove_null}{TRUE in default to reduce data}

\item{ss}{name of spreadsheet software. Error values from the given ss are replaced by NA}

\item{args1}{additional arguments given to first fread}

\item{args2}{additional arguments given to second fread}
}
\description{
transposing fread
}
