% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{freq_table}
\alias{freq_table}
\title{Frequencies Grouped by Another Column}
\usage{
freq_table(x, group_var, prop_var)
}
\arguments{
\item{x}{a data frame}

\item{group_var}{the variable to group by}

\item{prop_var}{the variable over which to calculate the frequencies}
}
\value{
A data frame with both the absolute and the relative distribution of `prop_var`
within `group_var`.
}
\description{
Obtain the relative distribution of a categorical variable within the group
member of a different variable.
}
\examples{
mtcars \%>\% freq_table(vs, cyl)
}
\seealso{
Inspired by SO question 24576515
}

