% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchRandomPeak.R
\name{generateRandomPeak}
\alias{generateRandomPeak}
\title{generateRandomPeak}
\usage{
generateRandomPeak(gtf, peakWidth = 70, seed = 111,
  intervalPerPeak = 1000)
}
\arguments{
\item{gtf}{GTF files}

\item{peakWidth}{mean width of the random peak}

\item{seed}{Seed for random number}

\item{intervalPerPeak}{average interval on transcript between each sampled peak}
}
