% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_genes2df.R
\name{genes2df}
\alias{genes2df}
\title{genes2df}
\usage{
genes2df(...)
}
\arguments{
\item{...}{The vector of genes.}
}
\value{
A data.frame
}
\description{
A function to convert multi-vectors to df
}
\author{
GM.W
}
