% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blackfly_timestamp_from_vidtime.R
\name{get_blackfly_timestamp_from_vidtime}
\alias{get_blackfly_timestamp_from_vidtime}
\title{Retrieves Blackfly Timestamp}
\usage{
get_blackfly_timestamp_from_vidtime(hab, cam, video_col, seconds_col)
}
\arguments{
\item{hab}{habitat spreadsheet}

\item{cam}{table of blackfly camera data}

\item{video_col}{column of video number in hab}

\item{seconds_col}{column of seconds in hab}
}
\description{
Retrieves Blackfly Timestamp from approximate video time
}
