% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blackfly_timestamp_from_vidtime.R
\name{get_blackfly_timestamp_from_vidtime_1}
\alias{get_blackfly_timestamp_from_vidtime_1}
\title{Returns Blackfly Timestamp based on Video Number and Seconds in}
\usage{
get_blackfly_timestamp_from_vidtime_1(video_col, seconds_col, cam_timestamp,
  cam_filepath)
}
\arguments{
\item{video_col}{vector of video numbers (must be same length as seconds_col)}

\item{seconds_col}{vector of seconds in video (must be same length as video_col)}

\item{cam_timestamp}{vector of camera timestamps (must be same length as cam_filepath)}

\item{cam_filepath}{vector of filepaths from balckfly camera (must be same length as cam_timestamp)}
}
\description{
Frame number is grabbed from gigegrab and adjusted based on gigegrab number in first frame
}
