% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_HypackPos.R
\name{get_HypackPos}
\alias{get_HypackPos}
\title{Get ship position from Hypack log}
\usage{
get_HypackPos(Hypack)
}
\arguments{
\item{Hypack}{output of read_hypack}
}
\description{
Get ship position from Hypack log
}
