% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_HypackSpeed.R
\name{get_HypackSpeed}
\alias{get_HypackSpeed}
\title{Get speed from Hypack log}
\usage{
get_HypackSpeed(Hypack)
}
\arguments{
\item{Hypack}{output of read_hypack}
}
\description{
Get speed from Hypack log
}
