% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_data.R
\name{get_info_data}
\alias{get_info_data}
\title{This function gives the info on the dataframe}
\usage{
get_info_data(data)
}
\description{
This function gives the info on the dataframe
}
\examples{
get_info_data(data = iris)
}
