% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_gg_density.R
\name{gg_density}
\alias{gg_density}
\title{Create Simplified Density Plots with ggplot2}
\usage{
gg_density(
  data,
  var,
  title = "Densityplot",
  x_lim = c(-50, 200),
  x_label = NULL,
  y_label = NULL
)
}
\arguments{
\item{data}{The data frame containing the variable to plot.}

\item{var}{A string specifying the variable to plot.}

\item{title}{A string specifying the title of the plot. Default is 'Densityplot'.}

\item{x_lim}{A numeric vector specifying the x-axis limits. Default is c(-50, 200).}

\item{x_label}{A string specifying the label for the x-axis. Default is NULL.}

\item{y_label}{A string specifying the label for the y-axis. Default is NULL.}
}
\value{
A ggplot object, with aesthetic mappings and other graphical parameters pre-configured
for ease of use and visual appeal.
}
\description{
This function is designed to offer a middle ground between the often verbose syntax of ggplot2 and
the speed but limited customization of base R plotting functions. It allows for the creation of
visually appealing density plots with minimal syntax.
}
\note{
This function uses `aes_string()`, which internally evaluates the string provided.
Be cautious when using this function in an environment where the data or variable names
are not trusted.
}
\examples{
\dontrun{
data(mtcars)
gg_density(mtcars, var = "mpg")
}

}
\seealso{
For more advanced customizations, see the ggplot2 package documentation.
}
