% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_gg_density_gr.R
\name{gg_density_gr}
\alias{gg_density_gr}
\title{Create Grouped Density Plots with Minimal ggplot2 Syntax}
\usage{
gg_density_gr(
  data,
  var,
  group_var,
  title = "Densityplot",
  x_lim = c(-50, 200),
  x_label = NULL,
  y_label = NULL
)
}
\arguments{
\item{data}{The data frame containing the variables to plot.}

\item{var}{A string specifying the variable for density plotting.}

\item{group_var}{A string specifying the variable to group by.}

\item{title}{A string specifying the title of the plot. Default is 'Densityplot'.}

\item{x_lim}{A numeric vector specifying the x-axis limits. Default is c(-50, 200).}

\item{x_label}{A string specifying the label for the x-axis. Default is NULL.}

\item{y_label}{A string specifying the label for the y-axis. Default is NULL.}
}
\value{
A ggplot object, with aesthetic mappings and other graphical parameters pre-configured
for ease of use and visual appeal.
}
\description{
This function serves as a compromise between the often verbose syntax required by ggplot2 and
the quick but less visually customizable base R plotting. It enables the creation of grouped
density plots with minimal input.
}
\note{
This function uses `aes_string()`, which internally evaluates the string provided.
Be cautious when using this function in an environment where the data or variable names
are not trusted.
}
\examples{
\dontrun{
data(mtcars)
gg_density_gr(mtcars, var = "mpg", group_var = "am")
}

}
\seealso{
For more advanced customizations, see the ggplot2 package documentation.
}
