% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_gg_histogram.R
\name{gg_histogram}
\alias{gg_histogram}
\title{Create Histograms with Simplified ggplot2 Syntax}
\usage{
gg_histogram(
  data,
  var,
  title = "Histogram",
  x_lim = c(-50, 200),
  x_label = NULL,
  y_label = NULL,
  bins = 30
)
}
\arguments{
\item{data}{The data frame containing the variable to plot.}

\item{var}{A string specifying the variable for the histogram.}

\item{title}{A string specifying the title of the plot. Default is 'Histogram'.}

\item{x_lim}{A numeric vector specifying the x-axis limits. Default is c(-50, 200).}

\item{x_label}{A string specifying the label for the x-axis. Default is NULL.}

\item{y_label}{A string specifying the label for the y-axis. Default is NULL.}

\item{bins}{An integer specifying the number of bins in the histogram. Default is 30.}
}
\value{
A ggplot object, with aesthetic mappings and other graphical parameters pre-configured
for ease of use and visual appeal.
}
\description{
This function aims to offer a middle ground between the verbose syntax of ggplot2 and
the quick but less visually pleasing plots generated by R's internal plotting functions.
It allows for the creation of visually appealing histograms with minimal input.
}
\note{
This function uses `aes_string()`, which internally evaluates the string provided.
Be cautious when using this function in an environment where the data or variable names
are not trusted.
}
\examples{
\dontrun{
data(mtcars)
gg_histogram(mtcars, var = "mpg")
}

}
\seealso{
For more advanced customizations, refer to the ggplot2 package documentation.
}
