% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_gg_qq_plot.R
\name{gg_qq_plot}
\alias{gg_qq_plot}
\title{Create a QQ-Plot using ggplot2}
\usage{
gg_qq_plot(data, column = NULL, color = "steelblue", language = "english")
}
\arguments{
\item{data}{A dataframe containing the data.}

\item{column}{The name of the column for which the QQ-Plot should be created. Default is NULL.}

\item{language}{The language for the axis labels. Default is "english". Can be set to "german" for German labels.}
}
\value{
A ggplot2 object representing the QQ-Plot.
}
\description{
This function creates a QQ-Plot for a specific column in a dataframe.
You can choose the language of the axis labels between English and German.
}
\examples{
\dontrun{
gg_qq_plot(mtcars, column = "mpg", language = "english")
}
}
