% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbreaks.R
\name{ggbreaks}
\alias{ggbreaks}
\title{Custom ggplot2 Breaks}
\usage{
ggbreaks(x = "continuous", y = "nothing")
}
\arguments{
\item{x}{Character, indicating if the x-axis should be treated as "continuous".}

\item{y}{Character, indicating additional options for the y-axis, with "nothing" as default,
allowing for automatic calculation of breaks and minor breaks, or formatted labels.}
}
\value{
A list of ggplot2 theme and scale adjustments.
}
\description{
Adjusts the breaks for ggplot2 axis scales, with specific behavior based on the axis type.
For y-axes, it dynamically calculates breaks and minor breaks, optionally formatting the labels as percentages.
}
\examples{
ggbreaks(x="continuous", y="nothing")
}
