% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglayout.R
\name{gglayout}
\alias{gglayout}
\title{evaluate a string as a R code}
\usage{
gglayout(plots, pos_row = NULL, pos_col = NULL)
}
\arguments{
\item{plots}{list of plots made by ggplot2}

\item{pos_row}{row positions of each plots}

\item{pos_col}{col positions of each plots}
}
\description{
evaluate a string as a R code
}
