% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOKEGG.R
\name{GOKEGG}
\alias{GOKEGG}
\title{GO and KEGG analysis, with plot and tsv output}
\usage{
GOKEGG(
  genelist,
  IDtype = "SYMBOL",
  species = "hs",
  ont = "BP",
  outdir = "./",
  data.name = NULL,
  color = NULL,
  n = 10
)
}
\arguments{
\item{genelist}{Gene list to perform enrichment analysis on}

\item{IDtype}{GeneID type: c("ENSEMBL", "ENTREZID", "SYMBOL", "UNIPROT"), default is "SYMBOL"}

\item{species}{c("hs" or "mm"), default is "hs"}

\item{ont}{ontology of GO: c("BP", "MF", "CC", "ALL"), default is "BP"}

\item{outdir}{directory to place the enrichment *.tsv file and plot file, default is the work directory}

\item{data.name}{Gene list name, serving as prefix of output file, default is NULL}

\item{color}{plot color pass to myGOhist()}

\item{n}{number of term to plot, pass to myGOhist()}
}
\description{
GO and KEGG analysis, with plot and tsv output
}
\details{
Gene list will first converted using bitr(), then perform GO and KEGG pathway enrichment, output the enrichment terms
in .tsv file and hist plot in .pdf
}
