% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headline.R
\name{headline}
\alias{headline}
\title{Print Headline}
\usage{
headline(text, n_signs = 70, style = "=")
}
\arguments{
\item{text}{Text der Überschrift.}

\item{n_signs}{Breite der Überschrift.}

\item{style}{Zeichen, aus dem die Überschrift bestehen soll.}
}
\description{
Diese Funktion erzeugt eine Überschrift in der Konsole.
Das ist z.B. in Schleifen zwischen der Ausgabe von Eregbnissen gelegentlich praktisch.
}
\examples{
headline('Results of calculations', n_signs = 65, style = '-')
}
