% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_highcorrpairs.R
\name{highcorrpairs}
\alias{highcorrpairs}
\title{Identify Highly Correlated Pairs of Features}
\usage{
highcorrpairs(df, threshold = 0.7, exclude_features = NULL)
}
\arguments{
\item{df}{A data frame containing the features to be analyzed. The data frame should contain at least one numeric feature.}

\item{threshold}{A numeric value between 0 and 1 that sets the correlation threshold. Default is 0.7.}

\item{exclude_features}{A character vector specifying the names of features to be excluded from the analysis. Default is NULL.}
}
\value{
A data frame containing pairs of highly correlated features and their Pearson's correlation coefficients, sorted in descending order.
}
\description{
This function takes a data frame and identifies pairs of features that are highly correlated,
based on a specified correlation threshold.
}
\examples{
\dontrun{
df <- data.frame(x1 = rnorm(100), x2 = rnorm(100), x3 = rnorm(100))
df$x4 <- df$x1 + df$x2
highcorrpairs(df)
}

}
