% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_cramersV.R
\name{highest_cramersV}
\alias{highest_cramersV}
\title{Calculate and Filter Variables Based on Cramér's V with a Target Variable}
\usage{
highest_cramersV(
  dat,
  target,
  threshold = 0.5,
  exclude = NULL,
  ignore_chisq_warnings = TRUE,
  print_headline = FALSE
)
}
\arguments{
\item{dat}{A data frame containing the variables to be analyzed. Must be a data frame.}

\item{target}{The name of the target variable in the data frame. Must be a character string.}

\item{threshold}{The threshold for Cramér's V to filter variables. Default is 0.5. Must be a numeric value.}

\item{exclude}{A vector of variable names to exclude from the analysis. Default is NULL. Must be a character vector or NULL.}

\item{ignore_chisq_warnings}{Logical flag to suppress chi-squared warnings. Default is TRUE.}

\item{print_headline}{Logical flag to print a headline for the results using \code{mytools::headline()}. Default is FALSE.}
}
\value{
A data frame containing variables that have a Cramér's V greater than the specified threshold with the target variable.
The data frame has two columns: 'Variable' and 'cramersV'.
}
\description{
This function calculates Cramér's V, a measure of association between two nominal variables,
for each variable in a given data frame against a specified target variable.
It then filters the variables based on a given threshold for Cramér's V.
}
\examples{
\dontrun{
data <- data.frame(x1 = c(1, 2, 3), x2 = c(3, 4, 5), y = c(0, 1, 0))
result <- highest_cramersV(data, target = "y", threshold = 0.2)
}

}
\seealso{
\code{\link{cramersV}} for the function used to calculate Cramér's V.
}
