% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magick-image-ggplot.R
\name{image_ggplot2}
\alias{image_ggplot2}
\title{A faster version of magick::image_ggplot}
\usage{
image_ggplot2(image, interpolate = FALSE)
}
\arguments{
\item{image}{magick image object returned by \code{\link[=image_read]{image_read()}} or \code{\link[=image_graph]{image_graph()}}}

\item{interpolate}{If \code{TRUE} interpolate linearly, if \code{FALSE}
(the default) don't interpolate.}
}
\description{
A faster version of magick::image_ggplot
}
