% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_db_table.R
\name{import_db_table}
\alias{import_db_table}
\title{Load database table into memory as dataframe (tibble)}
\usage{
import_db_table(conn, table)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by dbConnect().}

\item{table}{a character string with the name of the remote table.}
}
\description{
Load database table into memory as dataframe (tibble)
}
