% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-include_graphics.R
\name{include_graphics2}
\alias{include_graphics2}
\title{A wrapper of knitr::include_graphics with file conversion}
\usage{
include_graphics2(path, auto_pdf = TRUE, auto_png = TRUE, dpi = NULL)
}
\arguments{
\item{path}{A character vector of image paths.}

\item{auto_pdf}{If knitting pdf and input path is svg file, use pdf file instead. Conversion is also performed if neccessary.}

\item{auto_png}{If knitting docx and input path is svg file, use png file instead. Conversion is also performed if neccessary.}

\item{dpi}{DPI (dots per inch) value. Used to calculate the output
width (in inches) of the images. This will be their actual width in pixels,
divided by \code{dpi}. If not provided, the chunk option \code{dpi} is used; if
\code{NA}, the output width will not be calculated.}
}
\description{
A wrapper of knitr::include_graphics with file conversion
}
