% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp.R
\name{interp}
\alias{interp}
\title{Interpolate Ship Position}
\usage{
interp(my_data)
}
\arguments{
\item{my_data}{column of ship's position data in Easting or Northing}
}
\description{
Interpolates the ship's approximately 2Hz data to 1Hz. Will only work if there are not Multiple NA's in a row and uniform intervals
}
\keyword{position}
