% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_keep_complete_cases.R
\name{keep_complete_cases}
\alias{keep_complete_cases}
\title{Handle Missing Data in a Dataset}
\usage{
keep_complete_cases(dataset, features = names(dataset))
}
\arguments{
\item{dataset}{A dataframe containing the data.}

\item{features}{Character vector specifying the features to consider for complete cases. Default is all features.}
}
\value{
A dataframe containing only complete cases.
}
\description{
This function keeps only the complete cases in the dataset for the specified features
and prints out information about the removed cases and the new dimensions.
}
\examples{
\dontrun{
new_data <- keep_complete_cases(old_data, features = c("feature1", "feature2"))
}
}
