% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepR.R
\name{keepR}
\alias{keepR}
\title{keepR Class}
\description{
A class for storing and retrieving various metrics for different models.
}
\examples{
\dontrun{
# Initialize the collector with metric names
collector <- keepR$new(c("R2", "Adj_R2", "RMSE"))

# Add a single model
collector$add("model1", c(R2 = 0.9, Adj_R2 = 0.89, RMSE = 0.1))

# Add multiple models in a batch
collector$add_batch(
  model_names = c("model2", "model3"),
  metrics_list = list(c(R2 = 0.85, Adj_R2 = 0.84, RMSE = 0.12), c(R2 = 0.8, Adj_R2 = 0.79, RMSE = 0.13))
)

# Retrieve metrics sorted by a specific metric
sorted_metrics <- collector$get_metrics("R2")

# Retrieve features of a specific model
features_model1 <- collector$get_features("model1")

# Remove a model
collector$remove("model3")

# Save metrics to a CSV file
collector$save_metrics("R2", "metrics.csv")

# Save metrics to a JSON file
collector$save_metrics("R2", "metrics.json")
}

## ------------------------------------------------
## Method `keepR$add_batch`
## ------------------------------------------------

\dontrun{
collector <- keepR$new(c("R2", "Adj_R2", "RMSE"))
collector$add_batch(
  model_names = c("model1", "model2"),
  metrics_list = list(c(R2 = 0.9, Adj_R2 = 0.89, RMSE = 0.1), c(R2 = 0.85, Adj_R2 = 0.84, RMSE = 0.12))
)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{models}}{A list to store models along with their metrics.}

\item{\code{metric_names}}{A vector of names for the metrics to be stored.}

\item{\code{log_file}}{Optional. Name of a logfile. If nothing passed, log is printed in console}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-keepR-new}{\code{keepR$new()}}
\item \href{#method-keepR-add}{\code{keepR$add()}}
\item \href{#method-keepR-add_batch}{\code{keepR$add_batch()}}
\item \href{#method-keepR-get_features}{\code{keepR$get_features()}}
\item \href{#method-keepR-remove}{\code{keepR$remove()}}
\item \href{#method-keepR-get_metrics}{\code{keepR$get_metrics()}}
\item \href{#method-keepR-save_metrics}{\code{keepR$save_metrics()}}
\item \href{#method-keepR-clone}{\code{keepR$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keepR-new"></a>}}
\if{latex}{\out{\hypertarget{method-keepR-new}{}}}
\subsection{Method \code{new()}}{
Initializes an empty list to store models and sets the metric names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keepR$new(metric_names, log_file = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metric_names}}{A vector of names for the metrics to be stored.}

\item{\code{log_file}}{Optional. A string specifying the file path where log entries will be saved.
If provided, log entries will be written to this file in CSV format (separated by ',').
If not provided, log entries will be printed to the console.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keepR-add"></a>}}
\if{latex}{\out{\hypertarget{method-keepR-add}{}}}
\subsection{Method \code{add()}}{
Adds a model along with its metrics. Keep in mind: Adding a model that alreadys exists will be overwritten.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keepR$add(model_name, metrics, features = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{The name of the model.}

\item{\code{metrics}}{A named vector or list containing the metrics.}

\item{\code{features}}{A named vector or list containing the metrics.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keepR-add_batch"></a>}}
\if{latex}{\out{\hypertarget{method-keepR-add_batch}{}}}
\subsection{Method \code{add_batch()}}{
Adds multiple models along with their metrics in a batch.Keep in mind: Adding a model that alreadys exists will be overwritten.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keepR$add_batch(model_names, metrics_list, features_list = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_names}}{A vector of names for the models.}

\item{\code{metrics_list}}{A list of named vectors or lists containing the metrics for each model.}

\item{\code{features_list}}{An optional list of named vectors or lists containing the features for each model.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
collector <- keepR$new(c("R2", "Adj_R2", "RMSE"))
collector$add_batch(
  model_names = c("model1", "model2"),
  metrics_list = list(c(R2 = 0.9, Adj_R2 = 0.89, RMSE = 0.1), c(R2 = 0.85, Adj_R2 = 0.84, RMSE = 0.12))
)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keepR-get_features"></a>}}
\if{latex}{\out{\hypertarget{method-keepR-get_features}{}}}
\subsection{Method \code{get_features()}}{
Retrieves the features used in a specific model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keepR$get_features(model_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{The name of the model.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector containing the features used in the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keepR-remove"></a>}}
\if{latex}{\out{\hypertarget{method-keepR-remove}{}}}
\subsection{Method \code{remove()}}{
Removes a model based on its name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keepR$remove(model_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{The name of the model to remove.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keepR-get_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-keepR-get_metrics}{}}}
\subsection{Method \code{get_metrics()}}{
Sorts and displays the models based on a specific metric.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keepR$get_metrics(sort_by, as_dataframe = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sort_by}}{The name of the metric to sort by.}

\item{\code{as_dataframe}}{Logical, whether to return the result as a data frame. Default is TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame or printed output based on the `as_dataframe` parameter.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keepR-save_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-keepR-save_metrics}{}}}
\subsection{Method \code{save_metrics()}}{
Save metrics as csv or json based on the file extension
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keepR$save_metrics(sort_by, file_name, subdomain = "results", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sort_by}}{output sorting variable}

\item{\code{file_name}}{Filename, no path, with extension (.csv or .json)}

\item{\code{subdomain}}{subdomain where file shall be saved. If subdoomain does not exist, it will be created using mytools::create_directories(subdomain). Default is 'results'}

\item{\code{...}}{further arguments to mytools::write.csv.fast or toJSON}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL - saving as side effect
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keepR-clone"></a>}}
\if{latex}{\out{\hypertarget{method-keepR-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keepR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
