% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansp2.R
\name{kmeansp2}
\alias{kmeansp2}
\title{kmeans++ from http://mahito.info/kmeansp2.html}
\usage{
kmeansp2(x, k, iter.max = 10, nstart = 1, ...)
}
\arguments{
\item{x}{numeric matrix of data, or an object that can be coerced to such a matrix (such as a numeric vector or a data frame with all numeric columns)}

\item{k}{number of clusters}

\item{iter.max}{the maximum number of iterations allowed}

\item{nstart}{if centers is a number, how many random sets should be chosen?}

\item{...}{additional arguments given to kmeans()}
}
\description{
kmeans++ from http://mahito.info/kmeansp2.html
}
