% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansp3.R
\name{kmeansp3}
\alias{kmeansp3}
\title{kmeans+++. You can specify some centers meaningly and others randomly based on the method of kmeans++.}
\usage{
kmeansp3(x, k = 2, centers_manual = NULL, scale = TRUE,
  iter.max = 1000, nstart = 1, algorithm = "Lloyd")
}
\arguments{
\item{x}{input data}

\item{k}{number of clusters}

\item{centers_manual}{partial or all selection of initial centers}

\item{scale}{TRUE or FALSE to scale x}

\item{iter.max}{the maximum number of iterations allowed.}

\item{nstart}{if centers is a number, how many random sets should be chosen?}

\item{algorithm}{default to Lloyd}
}
\description{
kmeans+++. You can specify some centers meaningly and others randomly based on the method of kmeans++.
}
