% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatexFromDf.R
\name{LatexFromDf}
\alias{LatexFromDf}
\title{LatexFromDf}
\usage{
LatexFromDf(df, path)
}
\arguments{
\item{df}{A DataFrame.}

\item{path}{The path where to save the new pdf.}
}
\value{
Nothing - but create a pdf doc with the table in the righ path
}
\description{
Transform and save a dataframe as a Latex Pdf and a csv copy in anycase. save pdf in the same doc.
}
\examples{
data_summary(c(1, 2, 3, 4, 5))
}
