% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatexFromModels.R
\name{LatexFromModels}
\alias{LatexFromModels}
\title{LatexFromModels}
\usage{
LatexFromModels(models, path, modelnames = NULL, title = NULL)
}
\arguments{
\item{models}{A model or a list of modelt}

\item{path}{The path where to save the new pdf.}

\item{modelnames}{The names of the models}

\item{title}{The title of the table}
}
\value{
Nothing - but create a pdf doc with the table in the right path
}
\description{
Transform and save a model - either linear or mixed lmer regression - as a Latex Pdf and a rdata copy in anycase. save pdf in the same doc.
}
\examples{
LatexFromModels(models, "./output/modeleffect1")
}
