% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{load_dataset}
\alias{load_dataset}
\title{Load Dataset from a Subdirectory}
\usage{
load_dataset(dataset, subdomain = "data", sep = ",")
}
\arguments{
\item{dataset}{The name of the dataset (CSV file) you want to load.}

\item{subdomain}{The subdirectory where the dataset is located, relative to the working directory. Default is 'data'.}

\item{sep}{The field separator character. Values on each line of the file are separated by this character. Default is ','.}
}
\value{
A dataframe containing the data read from the CSV file.
}
\description{
This function reads a CSV file from a specified subdirectory in the working directory.
By default, it looks in the 'data' subdirectory.
}
\examples{
\dontrun{
# Loading a dataset called "my_data.csv" from the "data" directory.
my_data <- load_dataset("my_data.csv")

# Loading a dataset called "another_data.csv" from a "custom_data" directory.
another_data <- load_dataset("another_data.csv", subdomain = "custom_data", sep = "\t")
}
}
