% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.r
\name{loader}
\alias{loader}
\title{Data frame loader}
\usage{
loader()
}
\value{
A data frame read in via read.csv with column headings taken from 
the first row of the file.
}
\description{
Loads a csv file via read.csv into a data frame with headers. 
Saves a few characters typing, so worth it in my book.
}

