% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtest.R
\name{lrtest}
\alias{lrtest}
\title{This function is used to conduct the likelihood ratio test.}
\usage{
lrtest(dat, p)
}
\arguments{
\item{dat}{list of values being tested}

\item{p}{hypothesis probability}
}
\value{
list
}
\description{
Given a vector of 0s and 1s indicate failure and success,
you can test whether the probability of success equals
the hypothesised one.
Test staitstic = 2[y*log(y/p) + (n-y)log((n-y)/(n - n*p))]
See http://www.biostat.umn.edu/~dipankar/bmtry711.11/lecture_02.pdf
}
\examples{
#' lrtest(rnorm(100),0.01)
}

