% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2png.R
\name{map2png}
\alias{map2png}
\alias{map2png.default}
\alias{map2png.qm_xmap}
\title{Convert qm_xmap objects to png}
\usage{
map2png(x, nm = "map2png.png", .min = NA, .max = NA,
  lut = c("viridis", "gray"), ...)

\method{map2png}{default}(x, nm = "map2png", .min = NA, .max = NA,
  lut = c("viridis", "gray"), ...)

\method{map2png}{qm_xmap}(x, nm = "map2png", .min = NA, .max = NA,
  lut = "viridis", ...)
}
\arguments{
\item{x}{A matrix-like numeric object}

\item{nm}{A string to save `x` as png file.}

\item{.min, .max}{Minimum and maximum values for range of `scales::squish()`}

\item{lut}{Look up tables (viridis or gray). Partial matching is allowed.}

\item{...}{Discarded.}
}
\description{
Convert qm_xmap objects to png
}
