% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchRandomPeak.R
\name{matchRandomPeak}
\alias{matchRandomPeak}
\title{matchRandomPeak}
\usage{
matchRandomPeak(peak.gr, randomPeaks, txdb, nPeaks = 1e+10)
}
\arguments{
\item{peak.gr}{The GRange object of peak}

\item{randomPeaks}{The GRange object of random Peaks}

\item{txdb}{The TxDb object for the annotation}
}
