% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_McFaddenR2.R
\name{McFaddenR2}
\alias{McFaddenR2}
\title{Compute McFadden's Pseudo R-squared}
\usage{
McFaddenR2(model, verbose = T)
}
\arguments{
\item{model}{A fitted model object for which `logLik()` and `update()` methods exist.}

\item{verbose}{Logical. If `TRUE`, the function prints McFadden's Pseudo R-squared to the console. If `FALSE`, it returns the value silently.}
}
\value{
Numeric value representing McFadden's Pseudo R-squared, or NULL if it could not be computed.
}
\description{
This function calculates McFadden's Pseudo R-squared for generalized linear models.
It is designed primarily with `clm()` models in mind but is coded to be generalizable to other model types that support `logLik()` and `update()` functions.
}
\examples{
\dontrun{
library(ordinal)
fit <- clm(rating ~ temp + contact, data = wine)
McFaddenR2(fit)
}

}
