% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_cvision_csv.R
\name{merge_cvision_csv}
\alias{merge_cvision_csv}
\title{Merge C-Vision CSV's}
\usage{
merge_cvision_csv(file_list = list.files(pattern = "\\\\.csv$"),
  frames_per_sec, vid_length)
}
\arguments{
\item{file_list}{list of cvision files to be merged}

\item{frame_per_sec}{frames per second}

\item{Video}{Length in Minutes}
}
\description{
Merges all C-Vision CSV's into 1 CSV using a list of files or will merge all csv's in current working directory. NOTE: Use total frame for blackfly/AVT and Total Time for axis when trying to match to timestamp or position
}
