% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_modus.R
\name{modus}
\alias{modus}
\title{Calculate the mode of a numeric or integer vector}
\usage{
modus(x)
}
\arguments{
\item{x}{A numeric or integer vector}
}
\value{
The mode of the vector. Returns a vector of modes if the data is multimodal.
}
\description{
This function calculates the mode of a given numeric or integer vector.
If the data is multimodal, the function returns a vector containing all modes.
}
\examples{
modus(c(1, 2, 2, 3, 3, 4))
modus(c(1, 1, 2, 2, 3))
}
