% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_Cox.R
\name{Multi_Cox}
\alias{Multi_Cox}
\title{Multi_Cox}
\usage{
Multi_Cox(factor_list = factor_list, 
data_surv = data_surv,
time = "OS.time",
event = "OS.event")
}
\arguments{
\item{factor_list}{the factor of interset}

\item{data_surv}{the df containing factor_list, time, & event}
}
\value{
A data.frame contains \code{p.value}, \code{HR}, \code{HR.confint.lower}, \code{HR.confint.upper}, \code{HR.CI}.
}
\description{
A function to conduct multivariate Cox regression
}
\author{
GM.W
}
