% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_slope.R
\name{multi_slope}
\alias{multi_slope}
\title{Calculates multiscale slope of bathymetry/DEM
DOES NOT GIVE SIMILAR VALUES TO QGIS. NEED TO CHECK!!!!!!!!!!!!
Calculates multiscale slope of bathymetry/DEM in degrees. Any window containing an NA will result in an NA for slope. This function uses the Wood (1996) method which fits a bivariate quadratic equation (Evans 1980) and to an n x n grid and then calculates slope using ordinary least squares. For more details see Multiscale Terrain Analysis of Multibeam Bathymetry Data for Habitat Mapping on the Continental Slope by Wilson ey al 2007, pages 6-9.}
\usage{
multi_slope(bathy, window_size = 3)
}
\arguments{
\item{bathy}{bathymetry or DEM raster}

\item{window_size}{odd integer specifying dimensions of window (default is 3x3)}
}
\description{
Calculates multiscale slope of bathymetry/DEM
DOES NOT GIVE SIMILAR VALUES TO QGIS. NEED TO CHECK!!!!!!!!!!!!
Calculates multiscale slope of bathymetry/DEM in degrees. Any window containing an NA will result in an NA for slope. This function uses the Wood (1996) method which fits a bivariate quadratic equation (Evans 1980) and to an n x n grid and then calculates slope using ordinary least squares. For more details see Multiscale Terrain Analysis of Multibeam Bathymetry Data for Habitat Mapping on the Continental Slope by Wilson ey al 2007, pages 6-9.
}
