% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_multiple_ad_test.R
\name{multiple.ad.test}
\alias{multiple.ad.test}
\title{Conduct Anderson-Darling Tests for Normality on Multiple Variables}
\usage{
multiple.ad.test(dataset, variables, ...)
}
\arguments{
\item{dataset}{A data frame containing the variables to be tested.}

\item{variables}{A vector of character strings specifying the variable names to be tested.}

\item{...}{Additional arguments passed on to `nortest::ad.test`.}
}
\value{
A data frame containing the variable names, p-values, and significance levels of the Anderson-Darling tests.
}
\description{
This function performs Anderson-Darling tests for normality on a list of user-specified variables in a given dataset.
One of the advantages of this function is its ability to run tests for multiple variables in a loop, compiling the results into a single data frame for easy interpretation.
}
\examples{
\dontrun{
data <- data.frame(x = rnorm(100), y = rnorm(100))
multiple.ad.test(data, c("x", "y"))
}

}
