% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_multiple.wilcox.test.R
\name{multiple.wilcox.test}
\alias{multiple.wilcox.test}
\title{Compare groups Using the Mann-Whitney U Test}
\usage{
multiple.wilcox.test(dataset, features, group_var, exact = FALSE, ...)
}
\arguments{
\item{dataset}{A data frame containing the variables to be tested.}

\item{features}{A vector of character strings specifying the variable names to be tested.}

\item{group_var}{The name of the variable that identifies the groups.}

\item{exact}{Logical, indicating if an exact p-value should be computed.}

\item{...}{Additional arguments passed on to `wilcox.test`.}
}
\value{
A data frame containing the variable names, p-values, medians, and means for each group,
and significance levels.
}
\description{
This function performs the Mann-Whitney U test on a list of specified variables
to compare two groups within a given dataset.
group 1 comprises observations where the grouping variable equals 0,
and group 2 comprises observations where the grouping variable equals 1.
}
\examples{
\dontrun{
data <- data.frame(x = rnorm(100), y = rnorm(100), group = sample(0:1, 100, replace = TRUE))
multiple.wilcox.test(data, c("x", "y"), 'group')
}
}
