% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myGOhist}
\alias{myGOhist}
\title{GO hist plot}
\usage{
myGOhist(data, n = 10, color = NULL)
}
\arguments{
\item{data}{data.frame of enrichment result from ClusterProfiler enrichXxx family function}

\item{n}{number of terms to plot, default is 10}

\item{color}{color of the hist plot}
}
\value{
a ggplot object
}
\description{
GO hist plot
}
\details{
GO hist plot
}
