% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytools-package.r
\docType{package}
\name{mytools}
\alias{mytools}
\alias{mytools-package}
\title{This package collects my commonly used packages}
\description{
It includes a loglikelihood ratio test,
a function to find chibar -- the coefficient of tail independence
as well as a function to find kappa the tail index
and a function for piecewise linear scoring fucntion.
}

