% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_vert.R
\name{n_vert}
\alias{n_vert}
\title{Counts number of vertices in a spatial object}
\usage{
n_vert(shp)
}
\arguments{
\item{shp}{supported sp object}
}
\description{
Counts number of vertices in a spatial object. Supported classes include SpatialPolygons, SpatialPolygonsDataFrame, SpatialLines, SpatialLinesDataFrame, SpatialPoints, and SpatialPointsDataFrame.
}
