% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_time.R
\name{nearest_time}
\alias{nearest_time}
\title{Match to nearest time}
\usage{
nearest_time(time1, time2, unit = "secs", output_dif = FALSE,
  break_ties = "earlier")
}
\arguments{
\item{time1}{vector of timestamps (class of "POSIXct" "POSIXt")}

\item{time2}{vector of timestamps to campare time 1 to (class of "POSIXct" "POSIXt")}

\item{unit}{time unit ("secs", "mins", "hours", "days", or "weeks"): defualt is "secs"}

\item{output_dif}{Logical specifying whether or not the time difference for each row should be output (default is FALSE}

\item{break_ties}{if to times from time2 are equidistant should the "earlier" or "later" time be chosen (default is "earlier")}
}
\description{
Finds nearest time stamp in time 2 for each row of time 1. Optionally can include difference.
}
