% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_geochem.R
\name{norm_geochem}
\alias{norm_geochem}
\title{Normalize geochemical data and tydir::gather for spider diagram}
\usage{
norm_geochem(x, keep = NULL, unit = "ppm",
  std = "~/R/geochem/ms95.csv")
}
\arguments{
\item{x}{Input data.frame. Components not in the std file won't be normalized.}

\item{keep}{Components which should not be normalized.}

\item{unit}{Chose unit of input data}

\item{std}{Standard file for normalization.}
}
\description{
Normalize geochemical data and tydir::gather for spider diagram
}
