% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorthstarGGA2EK.R
\name{NorthstarGGA2EK}
\alias{NorthstarGGA2EK}
\title{Reformat postion from Northstar-941X--GGA.lab files to look like those from an EK cruise track csv}
\usage{
NorthstarGGA2EK(file_name)
}
\arguments{
\item{file_name}{path to Northstar-941X--GGA.lab file or Northstar-941X---GGA_*.Raw}
}
\description{
Reformat postion from Northstar-941X--GGA.lab files to look like those from an EK cruise track csv
}
