% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NotionCrap.R
\name{NotionCrap}
\alias{NotionCrap}
\title{NotionCrap}
\usage{
NotionCrap()
}
\arguments{
\item{folder_path}{the local path to the folder I want the link from}
}
\value{
A DataFrame with file names and their links
}
\description{
I forgot what this is for
}
\examples{
OneDriveLinks("/Users/leohenry/Library/CloudStorage/OneDrive-Personnel/Documents/Cours/Projets/CrimeUS_Fogglesong-Levy/2024_02_LA_DeclinationRates/output/240613-DRandCrime/3.240627-Changes/1-CityAnalysis")
}
