% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneDriveLinks.R
\name{OneDriveLinks}
\alias{OneDriveLinks}
\title{OneDriveLinks}
\usage{
OneDriveLinks(folder_path)
}
\arguments{
\item{folder_path}{the local path to the folder I want the link from}
}
\value{
A DataFrame with file names and their links
}
\description{
Gets the OneDrive links for the file in the folder in folder_path
}
\examples{
OneDriveLinks("/Users/leohenry/Library/CloudStorage/OneDrive-Personnel/Documents/Cours/Projets/CrimeUS_Fogglesong-Levy/2024_02_LA_DeclinationRates/output/240613-DRandCrime/3.240627-Changes/1-CityAnalysis")
}
