% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_bins.R
\name{order_bins}
\alias{order_bins}
\title{Order Percent Cover Bins and gives their corresponding frequency}
\usage{
order_bins(percent_cover)
}
\arguments{
\item{percent_cover}{a character vector of percent cover data in ranges formatted as "lower-upper"}
}
\description{
Returns a tibbe of ordered percent cover bins and their corresponding frequency. Percent Cover Bins are ordered in ascending order based on lower range. Ties are broken by upper range.
}
