% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_bins.R
\name{order_bins2}
\alias{order_bins2}
\title{Order Percent Cover Bins, gives their corresponding frequency, and lower and upper ranges}
\usage{
order_bins2(percent_cover)
}
\arguments{
\item{percent_cover}{a character vector of percent cover data in ranges formatted as "lower-upper"}
}
\description{
Returns a tibbe of ordered percent cover bins, their corresponding frequency, as well as lower and higher ranges of bins. Percent Cover Bins are ordered in ascending order based on lower range. Ties are broken by upper range.
}
