% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_Out_image&path.R
\name{out_image}
\alias{out_image}
\title{out_image}
\usage{
out_image(x=NULL,
filename = filename,
width = width,
height = height,
method = "export")
}
\arguments{
\item{x}{The plot. Default is \code{NULL} (\code{last_plot()}).}

\item{filename}{The name of the exporting figure.}

\item{width}{The width of figure.}

\item{height}{The height of figure.}

\item{method}{The method used to export the plot. Choose "export"(default) or "ggsave".}
}
\value{
\code{NULL}
}
\description{
Export plot using either "export" or "ggsave".
}
\author{
GM.W
}
