% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline_raster.R
\name{outline_raster}
\alias{outline_raster}
\title{Outlines non-na values of a raster}
\usage{
outline_raster(r, fact = 10)
}
\arguments{
\item{r}{A Raster object}

\item{fact}{aggregation factor for raster::aggregate. Useful if have very high resolution raster and tacing it causes memory issues. Also useful for filling "holes" in the middle of rasters where there may be some NA vals. By default fact=10}
}
\description{
Outlines non-na values of a raster. May want to follow up with rgeos::gsimplify to reduce number of vertices.
}
