% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs2.R
\name{pairs2}
\alias{pairs2}
\title{This function plots table for paired correlations with coefficients in upper
panel and scatterplots in lower panel}
\usage{
pairs2(df)
}
\arguments{
\item{df}{is dataset with numeric columns}
}
\value{
plot described above
}
\description{
This function plots table for paired correlations with coefficients in upper
panel and scatterplots in lower panel.
}
\examples{
df <- data.frame(1:10, 11:20, (1:10)^2)
pairs2(df)
}
\author{
Grag2015
}

