% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_video_fps.R
\name{parse_video_fps}
\alias{parse_video_fps}
\title{Gets frames per second from video info}
\usage{
parse_video_fps(vid_info)
}
\arguments{
\item{vid_info}{output from video_info function}
}
\description{
Parses fps of video from output of video_info function
}
