% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_video_time.R
\name{parse_video_time}
\alias{parse_video_time}
\title{Gets duration from video info}
\usage{
parse_video_time(vid_info)
}
\arguments{
\item{vid_info}{output from video_info function}
}
\description{
Parses duration of video in seconds from output of video_info function
}
