% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCC.R
\name{PCC}
\alias{PCC}
\title{Computes p-value of accuracy of a classifer according to the Proportional Chance Criterion}
\usage{
PCC(my_data, acc, iter)
}
\arguments{
\item{my_data}{a vector or one column dataframe/tibble}

\item{acc}{Accuracy of classifier}

\item{iter}{Number of iterations to run}
}
\description{
Computes p-value of accuracy of a classifer according to the Proportional Chance Criterion based on a specified number of iterations.
}
