% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perprim2perhard.R
\name{perprim2perhard}
\alias{perprim2perhard}
\title{Percent Primary to Percent Hardbottom}
\usage{
perprim2perhard(my_data, indur_col, per_col)
}
\arguments{
\item{my_data}{habitat spreadsheet}

\item{indur_col}{column number of induration}

\item{per_col}{column number of percent primary}
}
\description{
Converts Percent Primary to Percent Hardbottom
}
\keyword{percent_cover}
