% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_cell_cell_communications.R
\name{plot_CCI_ccdist}
\alias{plot_CCI_ccdist}
\title{plot_CCI_ccdist}
\usage{
plot_CCI_ccdist(st_dec_meta,
CCI_res,
celltype_sender,
celltype_receiver,
title = NULL,
color = NULL,
size = 1,
if_plot_others = TRUE,
if_plot_density = TRUE,
if_plot_edge = TRUE,
if_show_arrow = TRUE,
if_order_cell = TRUE,
legend.position = "bottom",
arrow_width = 0.02,
arrow_col = "#000000",
plot_cells = NULL)
}
\arguments{
\item{st_dec_meta}{The metadata whose row names are cells and contains cell, x, y, and celltype.}

\item{CCI_res}{The result of \code{find_CCI}, which contains data.frame including lrpair, tf, and cell_pair.}

\item{celltype_sender}{Character. The type of sender cells.}

\item{celltype_receiver}{Character. The type of receiver cells. Default is \code{NULL}.}

\item{title}{Character. The title of the plot. Default is \code{NULL}.}

\item{color}{Character. The color to be used. Default is \code{NULL}.}

\item{size}{Numeric. The size of the points. Default is 1.}

\item{if_plot_others}{Logical. Whether to plot other cell types? Default is \code{TRUE}.}

\item{if_plot_density}{Logical. Whether to plot the density of sender and receiver cells? Default is \code{TRUE}.}

\item{if_plot_edge}{Logical. Whether to plot segment between cell pairs? Default is \code{TRUE}.}

\item{if_show_arrow}{Logical. Whether to plot arrow from sender to receiver cells? Default is \code{TRUE}.}

\item{if_order_cell}{Logical. Whether to order cell before plotting? Default is \code{TRUE}.}

\item{legend.position}{Character or two-dimension vector. The position of legend. Default is "bottom".}

\item{arrow_width}{Numeric. The width of arrow. Default is 0.02.}

\item{arrow_col}{Character. The color of arrow. Default is "#000000" (black).}

\item{plot_cells}{Character. The specified cells to plot. Default is \code{NULL}.}
}
\value{
A ggplot object (when if_plot_density = FALSE) or ggExtraPlot object (when if_plot_density = TRUE).
}
\description{
Plot sender and receiver cells in a scatter plot using ggplot2.
}
\author{
GM.W
}
