% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_cell_cell_communications.R
\name{plot_CCI_lrpath}
\alias{plot_CCI_lrpath}
\title{plot_CCI_lrpath}
\usage{
plot_CCI_lrpath(
  lr_paths_filtered,
  CCI_res,
  celltype_sender,
  celltype_receiver,
  ligand,
  receptor,
  TFs = NULL,
  title = NULL,
  color_celltype = NULL,
  color_function = NULL,
  size = 5,
  edge_width = 0.1,
  legend.position = "right",
  max_hop = NULL,
  flip = FALSE,
  ReturnData = FALSE
)
}
\arguments{
\item{lr_paths_filtered}{The result of \code{filter_lr_paths}, which contains data.frames including lrpairs, and pathways.}

\item{CCI_res}{The result of \code{find_CCI}, which contains data.frames including lrpair, tf, and cell_pair.}

\item{celltype_sender}{Character. The type of sender cells.}

\item{celltype_receiver}{Character. The type of receiver cells.}

\item{ligand}{Character. The expressed ligand of celltype_sender cells.}

\item{receptor}{Character. The expressed receptor of celltype_receiver cells.}

\item{TFs}{Character. The expressed TF genes of celltype_receiver cells. Default is \code{NULL}.}

\item{title}{Character. The title of the plot. Default is \code{NULL}.}

\item{color_celltype}{Character. The color used for celltype (sender and receiver).}

\item{color_function}{Character. The color used for gene function (Ligand, Receptor, Mediator, TF, Target).}

\item{size}{Numeric. The size of point. Default is 5.}

\item{edge_width}{Numeric. The width of arrow. Default is 0.1.}

\item{legend.position}{Character or two-demension vector. The position of legend. Default is "right".}

\item{max_hop}{Numeric. The maximum hop to plot. Default is \code{NULL}.}

\item{flip}{Logical. Whether to flip the plot. Default is \code{FALSE}.}

\item{ReturnData}{Logical. Whether to return the lr_path data (for further visualization in cytoscape). Default is FALSE.}
}
\value{
A ggplot object.
}
\description{
Plot a network of Ligand, Receptor, Mediators, TFs, and Targets.
}
\author{
GM.W
}
