% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_cell_cell_communications.R
\name{plot_CCI_path2gene}
\alias{plot_CCI_path2gene}
\title{plot_CCI_path2gene}
\usage{
plot_CCI_path2gene(
  st_dec_data,
  CCI_res,
  pathways = NULL,
  species = "Human",
  celltype_sender,
  celltype_receiver,
  ligand,
  receptor,
  max_hop = NULL,
  min_gene_num = 5,
  pvalue = 0.05,
  geneRatio = NULL,
  color = NULL,
  size = 2,
  title = NULL,
  ReturnData = FALSE
)
}
\arguments{
\item{st_dec_data}{Deconvoluted Spatial matrix.}

\item{CCI_res}{The result of \code{find_CCI}, which contains data.frame including lrpair, tf, and cell_pair.}

\item{pathways}{Pathways with source and destination. Default is \code{SpaTalk::pathways}.}

\item{species}{Character. "Human" or "Mouse" or other specified. Default is "Human".}

\item{celltype_sender}{Character. The sender cell type.}

\item{celltype_receiver}{Character. The receiver cell type.}

\item{ligand}{Character. The expressed ligand of celltype_sender cells.}

\item{receptor}{Character. The expressed receptor of celltype_receiver cells.}

\item{max_hop}{Numeric. The maximum hop to plot. Default is \code{NULL}.}

\item{min_gene_num}{Numeric. The minimum number of genes in one pathways. Default is 5.}

\item{pvalue}{Numeric. The pvalue threshold of enrichment analysis. Default is 0.05.}

\item{geneRatio}{Numeric. The ratio threshold of enrichment analysis. Default is \code{NULL}.}

\item{color}{Character. The color to be used. Default is \code{NULL}.}

\item{size}{Numeric. The size of text. Default is 2.}

\item{title}{Character. The title of the plot. Default is \code{NULL}.}

\item{ReturnData}{Logical. Whether to return the enrichment result. Default is FALSE (sankey plot).}
}
\value{
A ggplot object (when \code{ReturnData} is FALSE) or a data.frame (when \code{ReturnData} is TRUE).
}
\description{
Draw a sankey plot containing pathways and genes.
}
\author{
GM.W
}
