% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_enrichment&GSEA.R
\name{plot_enrich}
\alias{plot_enrich}
\title{plot_enrich}
\usage{
plot_enrich(enrich_obj, 
type=c("dot","bar","bubble"),
enrich=c("GO","KEGG"),
facet=FALSE,
regulate=c("up","down","dys","wrap"),
max_term=18,
title=NULL,
color=NULL)
}
\arguments{
\item{enrich_obj}{The result of \code{enrich_analysis}.}

\item{type}{The type of plot.}

\item{enrich}{Either "GO" or "KEGG".}

\item{facet}{Logical.}

\item{max_term}{The maximum number of terms to draw.}

\item{title}{The title of plot.}

\item{color}{The color used in plot. Default is "Set1".}

\item{regulation}{The regulation of genes used in enrichment analysis.}
}
\value{
A ggplot object.
}
\description{
A function to visualize the results of GO and KEGG enrichment analysis.
}
\examples{
plot_enrich(enrich_obj = enrich_obj,type = "dot",
enrich="GO",regulate = "up",title = "GO.up")

}
\author{
GM.W
}
