% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_enrichment&GSEA.R
\name{plot_GSEA}
\alias{plot_GSEA}
\title{plot_GSEA}
\usage{
plot_GSEA(enrich_obj, 
type=c("GSEAcurve","bar","dot"),
geneSetID=NULL,
regulate=NULL, 
max_signatures=6, 
addGene=NULL, 
sig=c("GO","KEGG","HALLMARK","REACTOME","TF","miRNA"),
color=NULL, 
subplot=3, 
legend.position=NULL, 
title=NULL)
}
\arguments{
\item{enrich_obj}{The result of \code{enrich_analysis}.}

\item{type}{The type of plot.}

\item{geneSetID}{The id of geneset.}

\item{regulate}{The regulation of signatures.}

\item{max_signatures}{The maximum number of signatures.}

\item{addGene}{The gene of interest to add on graph.}

\item{sig}{The signature choiced to visualize}

\item{color}{The color used for curve.}

\item{subplot}{The number of plots to draw.}

\item{legend.position}{The position of legend.}

\item{title}{The title of plot.}
}
\value{
A plot.
}
\description{
A function to visualize the results of GO and KEGG enrichment analysis.
}
\author{
GM.W
}
