% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_regression_diagnostics.R
\name{plot_qq_residuals}
\alias{plot_qq_residuals}
\title{Functions for Graphical Regression Diagnostics}
\usage{
plot_qq_residuals(model)
}
\arguments{
\item{model}{A fitted model object.}

\item{save}{Logical. Whether to save the plots or not. Default is FALSE.}

\item{praefix}{A string prefix for saved plot filenames.}

\item{subdomain}{The subdirectory within the current working directory where the plot will be saved. Default is 'plots'.}

\item{n_breaks}{The number of breaks for the histogram. Default is 40.}
}
\value{
Various diagnostic plots for the given regression model.
}
\description{
This set of functions provides various plots for regression diagnostics.
It includes QQ-plots for residuals (`plot_qq_residuals`), residual vs fitted plots (`plot_residual_vs_fitted`),
histograms of residuals (`plot_histogram_residuals`), and scale-location plots (`plot_scale_location`).
Additionally, it provides an option to save these plots (`plot_regr_diagnostics`).
}
\section{plot_regr_diagnostics}{

Generates and optionally saves all diagnostic plots.
}

\section{plot_qq_residuals}{

Generates a QQ-plot for the residuals of the model.
}

\section{plot_residual_vs_fitted}{

Generates a residual vs fitted plot for the model.
}

\section{plot_histogram_residuals}{

Generates a histogram for the residuals of the model.
}

\section{plot_scale_location}{

Generates a scale-location plot for the model.
}

\examples{
\dontrun{
# Fit a linear model
fit <- lm(mpg ~ wt + hp, data = mtcars)
# Generate and save all diagnostic plots
plot_regr_diagnostics(fit, save = TRUE, praefix = "my_model")
}
}
