% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.r
\name{plotbar}
\alias{plotbar}
\title{Plot means and error bars for a single factor}
\usage{
plotbar(data, bar, factor, error)
}
\arguments{
\item{data}{The dataset containing the values to be plotted.}

\item{bar}{The column of the dataset representing the values to be averaged.}

\item{factor}{The column representing the grouping factor.}

\item{error}{The type of error value to be calculated. Options are 
"sd", "se" or "ci".}
}
\value{
A bar plot (dynamite plot) of means and error bars for a single 
factor.
}
\description{
Plots a standard 'mean and error bar figure for levels of 
a single factor. Factors represented as continuous variables must be 
converted to an actual factor (using as.factor) before use.
}

