% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.r
\name{plotbar2}
\alias{plotbar2}
\title{Plot means and error bars for two factors}
\usage{
plotbar2(data, bar, factor, legend, error)
}
\arguments{
\item{data}{The dataset containing the values to be plotted.}

\item{bar}{The column of the dataset representing the values to be averaged.}

\item{factor}{The column representing the grouping factor to be shown on 
the x-axis.}

\item{legend}{The column representing the grouping factor to be shown in 
the legend.}

\item{error}{The type of error value to be calculated. Options are 
"sd", "se" or "ci".}
}
\value{
A bar plot (dynamite plot) of means and error bars for two factors.
}
\description{
Plots a standard 'mean and error bar figure for levels of 
two factors. Factors represented as continuous variables must be 
converted to an actual factor (using as.factor) before use. First factor 
is represented as categories on the x-axis and the second as levels in 
the plot legend.
}

