% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExpressionDESeq2.R
\name{plotExpressionDESeq2}
\alias{plotExpressionDESeq2}
\title{plotExpressionDESeq2}
\usage{
plotExpressionDESeq2(dds, geneName)
}
\arguments{
\item{dds}{The DESeq2 object}

\item{geneName}{The character or vector of gene names}
}
