% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.r
\name{plotline}
\alias{plotline}
\title{Plot mean and error bar as line chart, for single factor}
\usage{
plotline(data, val, factor, error = "sd")
}
\arguments{
\item{data}{The dataset containing the values to be plotted.}

\item{val}{The column of the dataset representing the values to be averaged.}

\item{factor}{The column representing the grouping factor (categories on 
the x-axis.}

\item{error}{The type of error value to be calculated. Options are 
"sd", "se" or "ci". Defaults to "sd".}
}
\value{
A line plot of means and error bars for a single factor.
}
\description{
Plots a standard 'mean and error bar figure for levels of 
a single factor as a line plot. Factors represented as continuous variables 
must be converted to an actual factor (using as.factor) before use.
}

