% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.r
\name{plotline2}
\alias{plotline2}
\title{Plot mean and error bar as line chart, for two factors, first as axis, 
second as legend.}
\usage{
plotline2(data, val, factor, legend, error)
}
\arguments{
\item{data}{The dataset containing the values to be plotted.}

\item{val}{The column of the dataset representing the values to be averaged.}

\item{factor}{The column representing the grouping factor (categories on 
the x-axis.}

\item{legend}{The column representing the grouping factor to be shown in 
the legend.}

\item{error}{The type of error value to be calculated. Options are 
"sd", "se" or "ci". Defaults to "sd".}
}
\value{
A line plot of means and error bars for two factors.
}
\description{
Plots a standard 'mean and error bar figure for levels of 
two factors as a line plot. Factors represented as continuous variables 
must be converted to an actual factor (using as.factor) before use.  First 
factor is represented as categories on the x-axis and the second as levels 
in the plot legend.
}

