% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneCoverage.R
\name{plotVirusCovPairs}
\alias{plotVirusCovPairs}
\title{plotVirusCovPairs}
\usage{
plotVirusCovPairs(IP_BAM, INPUT_BAM, X, geneModel, libraryType = "opposite",
  center = mean, annotation, ZoomIn = NULL, hideStrand = "none")
}
\arguments{
\item{IP_BAM}{The bam files for IP samples}

\item{INPUT_BAM}{The bam files for INPUT samples}

\item{geneModel}{The GRange object storing positive and negative strand as gene model of the virus.}

\item{libraryType}{The specification of strandness of the RNA library. "opposite" for mRNA stranded library that sequenced the reverse complementary strand; "same" for samll RNA library that sequenced the same strand.}

\item{center}{The function used to calculate the center of multiple replicates. Can be mean or median.}

\item{annotation}{The data.frame specify the genes on the viral genome. Needs to have column "chr", "start", "end", "gene".}

\item{ZoomIn}{The range to plot the coverage.}

\item{hideStrand}{The strand to hide when plot. Default is none.}
}
