% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_as_list.R
\name{print_as_list}
\alias{print_as_list}
\title{Print a Character Vector as a List}
\usage{
print_as_list(char_vector)
}
\arguments{
\item{char_vector}{A character vector that you want to print as a list.}
}
\value{
Prints the elements of the character vector as a list to the console.
}
\description{
This function takes a character vector and prints each element
in a list format, where each item starts with a dash followed by a space.
}
\examples{
# Example usage
my_char_vector <- c("Apfel", "Birne", "Kirsche")
print_as_list(my_char_vector)

# Output:
# - Apfel
# - Birne
# - Kirsche
}
