% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paircor.R
\name{printcor}
\alias{printcor}
\title{calculates pairs coeff.correlation}
\usage{
printcor(df, level = 0.5)
}
\arguments{
\item{df}{is dataset with numeric columns}
\item{level}{is coef. corr. level to print result}
}
\value{
calculates pairs coeff.correlation
}
\description{
This function calculates pairs coeff.correlation and print pairs which have correlation greater than level

}

