% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quit.R
\name{q}
\alias{q}
\title{short cut to terminate R}
\usage{
q(save = "no", status = 0, runLast = TRUE)
}
\arguments{
\item{save}{default to no}

\item{status}{default to 0}

\item{runLast}{default to TRUE}
}
\description{
short cut to terminate R
}
