% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radians.R
\name{radians}
\alias{radians}
\title{Converts a measurement in degrees to radians}
\usage{
radians(deg)
}
\arguments{
\item{deg}{measurement in degrees}
}
\description{
Converts a measurement in degrees to radians
}
