% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_ranger_imp_plot.R
\name{ranger_imp_plot}
\alias{ranger_imp_plot}
\title{Create Variable Importance Plots for ranger Random Forest Models}
\usage{
ranger_imp_plot(ranger_model, top_vars = 10, col = "steelblue")
}
\arguments{
\item{ranger_model}{A ranger Random Forest model object. Note that the function does not
actually require the `ranger` package to be loaded, as it directly accesses the
`$variable.importance` attribute of the model object.}

\item{top_vars}{An integer specifying the number of top variables to display. Default is 10.}

\item{col}{A string specifying the color of the bars. Default is 'steelblue'.}
}
\value{
A ggplot object, pre-configured for visual appeal and emphasizing the most important
variables according to the model.
}
\description{
This function provides a simplified and specialized approach to visualize variable importance
from Random Forest models built using the `ranger` package. It aims to simplify the verbose
syntax often required for ggplot2, allowing for quick but visually appealing plots.
}
\note{
This function directly accesses the `$variable.importance` attribute of the ranger model object.
Be cautious when using this function in an environment where the model object is not trusted.
}
\examples{
\dontrun{
# Assuming `rf_model` is a ranger Random Forest model
ranger_imp_plot(rf_model)
}

}
\seealso{
For more advanced customizations and other ranger functionalities, refer to the ranger package documentation.
}
