% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDMV.R
\name{RDMV}
\alias{RDMV}
\title{Calculates relative difference from mean value from bathymetry raster}
\usage{
RDMV(bathy, window_size = 3)
}
\arguments{
\item{bathy}{bathymetry raster}

\item{window_size}{odd integer specifying dimensions of window (default is 3x3)}
}
\description{
Calculates relative difference from mean value from bathymetry raster. Rages from -1 to 1+ indicates a crest and - values indicating a depression. Calculated over a n x n window as (bathy - local_mean)/local_range. Depth is assumed to be negative values in calculation. If any values within window are NA, calculated value is NA.
(Lecours et al; 2017, Towards a framework for terrain attribute selection in environmental studies)
}
