% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel2.R
\name{read_excel2}
\alias{read_excel2}
\title{transposing read_excel}
\usage{
read_excel2(input, sheet = 1, transposing = TRUE, error_as_NA = TRUE,
  rm_blank_col = TRUE)
}
\arguments{
\item{input}{input data}

\item{sheet}{numeric or character to select sheet}

\item{transposing}{transpose data frame when rows are variables and cols are observations}

\item{error_as_NA}{turn error as NA (e.g., #DIV/0!)}

\item{rm_blank_col}{boolean to decide removing blank columns}
}
\description{
transposing read_excel
}
