% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hypack.R
\name{read_hypack}
\alias{read_hypack}
\title{Reads RAW Hypack files}
\usage{
read_hypack(filename)
}
\arguments{
\item{filename}{one or more RAW hypack files as a character vector}
}
\description{
Reads RAW Hypack log files into R
}
