% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hypack.R
\name{read_hypack_helper}
\alias{read_hypack_helper}
\title{helper function for read_hypack}
\usage{
read_hypack_helper(filename)
}
\arguments{
\item{filename}{filename of raw Hypack file}
}
\description{
helper function for read_hypack
}
