% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ods2.R
\name{read_ods2}
\alias{read_ods2}
\title{transposing read_ods}
\usage{
read_ods2(input, sheet = 1, error_as_NA = TRUE, rm_blank_col = TRUE)
}
\arguments{
\item{input}{input data}

\item{sheet}{numeric or character to select sheet}

\item{error_as_NA}{turn error as NA (e.g., #DIV/0!)}

\item{rm_blank_col}{boolean to decide removing blank columns}
}
\description{
transposing read_ods
}
