% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tbl_interactive.R
\name{read_tbl}
\alias{read_tbl}
\title{interactively read spreadsheet (xls, xlsx, ods, csv)}
\usage{
read_tbl(input = NULL, sheet = NULL, transposing = NULL)
}
\arguments{
\item{input}{leave NULL when interactive}

\item{sheet}{leave NULL when interactive}

\item{transposing}{leave NULL when interactive}
}
\description{
interactively read spreadsheet (xls, xlsx, ods, csv)
}
