% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-data-frame.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.data.frame}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{base}{\code{\link[base]{as.data.frame}}}
}}

