% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_or.R
\name{regex_or}
\alias{regex_or}
\title{produce regular expression of "or" (e.g., ^(abc)|(def)$)}
\usage{
regex_or(x)
}
\arguments{
\item{x}{a character vector}
}
\description{
produce regular expression of "or" (e.g., ^(abc)|(def)$)
}
