% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{make report by rmarkdown}
\usage{
report(Rmd = "report.Rmd", wd = NULL, output_format = "all",
  browse = TRUE)
}
\arguments{
\item{Rmd}{name of Rmd file}

\item{wd}{working directory}

\item{output_format}{all in default}

\item{browse}{whether or not to browse html file when it is in output}
}
\description{
make report by rmarkdown
}
