% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_RSEM_extract_transcriptome&clinical_data.R
\name{RESM_preview}
\alias{RESM_preview}
\title{RESM_preview}
\usage{
RESM_preview()
}
\arguments{
\item{prefix_read}{The path of Xena data "TcgaTargetGTEX_phenotype.txt.gz".}
}
\value{
A data.frame containing the type and number of diseases.
}
\description{
Cast a glance at the type of disease-samples.
}
\author{
GM.W
}
