% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev.comp.R
\name{revComp}
\alias{revComp}
\title{revComp}
\usage{
revComp(x, rev = TRUE)
}
\arguments{
\item{x}{The sequence to be revserve complement.}

\item{rev}{Logical option to choose whether reverse the complementary sequence.}
}
\description{
Get the (option: reverse) complementary sequence of the input sequence.
}
