% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPCA_cum_exp.R
\name{rPCA_cum_exp}
\alias{rPCA_cum_exp}
\title{Cumulative Variation Explained Raster PCA}
\usage{
rPCA_cum_exp(raster_pca)
}
\arguments{
\item{raster_pca}{output of rasterPCA function}
}
\description{
Calculates cumulative variation explained from rasterPCA
}
\keyword{rasterPCA}
