% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPCA_prop_exp.R
\name{rPCA_prop_exp}
\alias{rPCA_prop_exp}
\title{Proportion of Variation Explained by each layer in a Raster PCA}
\usage{
rPCA_prop_exp(raster_pca)
}
\arguments{
\item{raster_pca}{output of rasterPCA function}
}
\description{
Calculates the Proportion of Variation Explained by each layer in a Raster PCA
}
\keyword{rasterPCA}
