% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_RSEM_extract_transcriptome&clinical_data.R
\name{RSEM_extract_whole}
\alias{RSEM_extract_whole}
\title{RSEM_extract_whole}
\usage{
RSEM_extract_whole(project = "STAD",
keep_cate = "stomach")
}
\arguments{
\item{project}{The name of your project, which will be the prefix of the saved data.}

\item{keep_cate}{The type of disease. Use \code{RESM_preview()} to Cast a glance.}

\item{prefix_read}{The path your "TcgaTargetGTEX_phenotype.txt.gz" file stores.}

\item{prefix_out}{The path to save the extract data.}
}
\description{
Extract and save the data of selected disease type from Xena data.
}
\author{
GM.W
}
