% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_oncoPredict.R
\name{run_oncoPredict}
\alias{run_oncoPredict}
\title{run_oncoPredict}
\usage{
run_oncoPredict(
  ExprData,
  batchCorrect = "eb",
  drugs = NULL,
  powerTransformPhenotype = TRUE,
  removeLowVaryingGenes = 0.2,
  minNumSamples = 10,
  printOutput = TRUE,
  removeLowVaringGenesFrom = "rawData"
)
}
\arguments{
\item{ExprData}{A expression matrix.}

\item{batchCorrect}{The method of batch correction. Choices are "eb" (default) for ComBat, "qn" for quantiles normalization or "none" for no homogenization.}

\item{drugs}{The candidate drugs selected. Use \code{mytools::candidate_drugs} to pick some.}

\item{powerTransformPhenotype}{Should the phenotype be power transformed before we fit the regression model? Default to TRUE, set to FALSE if the phenotype is already known to be highly normal.}

\item{removeLowVaryingGenes}{What proportion of low varying genes should be removed? 20 percent be default.}

\item{minNumSamples}{How many training and test samples are required. Print an error if below this threshold.}

\item{printOutput}{Set to FALSE to supress output.}

\item{removeLowVaringGenesFrom}{Determine method to remove low varying genes. Options are 'homogenizeData' and 'rawData'.}
}
\value{
A data.frame containing the predicted drug sensitivity based on exprData.
}
\description{
A function to easily run oncoPredict
}
\author{
GM_W
}
