% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Savegrph.R
\name{Savegrph}
\alias{Savegrph}
\title{Save Graph to PDF}
\usage{
Savegrph(grph, path, width = 10, height = 8)
}
\arguments{
\item{grph}{The graphical object to save. Typically, this would be an object generated by plotting functions in R, such as those from `ggplot2` or base R plotting functions.}

\item{path}{The path where the PDF file should be saved, excluding the file extension. The '.pdf' extension is automatically appended to this path in the function.}

\item{width}{Optional. The width of the PDF document in inches. Default is 10.}

\item{height}{Optional. The height of the PDF document in inches. Default is 8.}
}
\value{
Invisibly returns NULL. The function's primary effect is the side-effect of creating a PDF file at the specified location containing the graphical representation of the `grph` object.
}
\description{
Saves a graphical object (e.g., a plot) to a PDF file at the specified path. This function is useful for exporting generated plots to PDF format for reports, presentations, or further analysis.
}
\examples{
# Assuming `plot_obj` is a plot object created using ggplot2 or base R plotting
Savegrph(plot_obj, "path/to/save/your_plot")
}
