% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapiro-by-group.R
\name{shapiro_by_group}
\alias{shapiro_by_group}
\title{shapiro_by_group - Teste de Shapiro-Wilk por grupo}
\usage{
shapiro_by_group(data, factor_var, num_var)
}
\arguments{
\item{data}{Um objeto data.frame contendo as variáveis a serem testadas.}

\item{factor_var}{O nome da variável categórica que define os grupos.}

\item{num_var}{O nome da variável numérica a ser testada.}
}
\value{
Um objeto data.frame contendo os resultados do teste de Shapiro-Wilk para cada grupo.
}
\description{
Esta função realiza o teste de Shapiro-Wilk em um conjunto de dados para cada grupo
definido por uma variável categórica especificada. A função retorna os resultados
de cada teste, incluindo o valor da estatística W, o valor-p e uma indicação se a
distribuição é normal ou não (com base em um nível de significância de 0,05).
}
\examples{
# criar um conjunto de dados simulado
set.seed(123)
notas_turma_a <- rnorm(30, mean = 7, sd = 1)
notas_turma_b <- rnorm(40, mean = 6.5, sd = 1.5)
turma <- rep(c("Turma A", "Turma B"), c(30, 40))
dados <- data.frame(Turma = turma, Notas = c(notas_turma_a, notas_turma_b))

# chamar a função shapiro_by_group para testar a normalidade das notas de cada turma
resultados_shapiro <- shapiro_by_group(dados, "Turma", "Notas")
print(resultados_shapiro)

}
