% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_species.R
\name{sort_species}
\alias{sort_species}
\title{Sort Species List}
\usage{
sort_species(species_list)
}
\arguments{
\item{species_list}{vector of unique species names}
}
\description{
Sorts species names alphabetically but puts spp first in their respective groups and No IDs last
}
