% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_survival_cut_plot.R
\name{survival_cut_plot}
\alias{survival_cut_plot}
\title{survival_cut_plot}
\usage{
survival_cut_plot(surv_data, 
variable = "NEIL3", method = "cutpoint",
   time = "OS.time",event = "OS",
   CI = T, title = "TCGA-PDAC")
}
\arguments{
\item{surv_data}{The data.frame containing time, event, and genes of interest}

\item{variable}{Gene or observation of interest}

\item{method}{Choose one from c("cutpoint","median"), default is "cutpoint".}

\item{time}{The name of time variable.}

\item{event}{The name of event variable.}

\item{palette}{The color schemes for the K-M curves. Default is c("#4dbbd5","#e64b35").}

\item{CI}{Logical. Whether to draw the Confidence interval. Default is \code{FALSE}.}

\item{title}{The title of the plot. Default is \code{NULL}.}
}
\value{
A ggplot object.
}
\description{
A function to draw K-M curve according to the median or best cutoff value
}
\examples{
survival_cut_plot(surv_data, variable = "NEIL3", method = "cutpoint",
   time = "OS.time",event = "OS",CI = T, title = "TCGA-PDAC")
   
}
\author{
GM.W
}
