% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_survival_cut_plot.R
\name{survival_cut_res}
\alias{survival_cut_res}
\title{survival_cut_res}
\usage{
survival_cut_res(surv_data, 
   variable = "NEIL3", 
   method = "cutpoint",
   time = "OS.time",
   event = "OS")
}
\arguments{
\item{surv_data}{The data.frame containing time, event, and genes of interest}

\item{variable}{Gene or observation of interest}

\item{method}{Choose one from c("cutpoint","median"), default is "cutpoint".}

\item{time}{The name of time variable.}

\item{event}{The name of event variable.}
}
\value{
A data.frame containing p-value of survival analysis.
}
\description{
A function to calculate the p-value of survival analysis.
}
\examples{
survival_cut_plot(surv_data, variable = "NEIL3", method = "cutpoint",
   time = "OS.time",event = "OS")
   
}
\author{
GM.W
}
