% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TASSE.R
\name{TASSE}
\alias{TASSE}
\title{Calculates TASSE terrain attributes with 3x3 window}
\usage{
TASSE(bathy, na.rm = FALSE, return_aspect = FALSE)
}
\arguments{
\item{bathy}{bathymetry (or dem) raster}

\item{na.rm}{whether to remove na's (default is FALSE)}

\item{return_aspect}{logical indicating whether to return aspect (default is FALSE)}
}
\description{
Calculates TASSE terrain attributes using a 3x3 window including relative deviation from mean value (rdmv), aspect, eastness, norhtnessm slope, standard deviation, and local mean. Depth is assumed to be negative values in calculation. By default if any values within window are NA, calculated value is NA. Slope and aspect are returned in degrees.
(Lecours et al; 2017, Towards a framework for terrain attribute selection in environmental studies)
}
