% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{test}
\alias{test}
\title{Execute \pkg{test_that} tests in a package.}
\usage{
test(pkg = ".", filter = NULL, ..., path = here::here())
}
\arguments{
\item{pkg}{package description, can be path or package name. See
\code{\link[=as.package]{as.package()}} for more information}

\item{filter}{If not \code{NULL}, only tests with file names matching this
regular expression will be executed.  Matching will take on the file
name after it has been stripped of \code{"test-"} and \code{".R"}.}

\item{...}{additional arguments passed to \code{\link[testthat:test_dir]{testthat::test_dir()}} and
\code{\link[covr:package_coverage]{covr::package_coverage()}}}
}
\description{
\code{test()} is a shortcut for \code{\link[testthat:test_dir]{testthat::test_dir()}}, it runs all of a
package's tests.
\code{test_file} runs \code{test()} on the active file.
\code{test_coverage()} computes test coverage for your package. It is a shortcut
for \code{\link[covr:package_coverage]{covr::package_coverage()}} and \code{\link[covr:report]{covr::report()}}.
\code{test_coverage_file()} computes test coverage for the active file. Is a
shortcut for \code{\link[covr:file_coverage]{covr::file_coverage()}} and \code{\link[covr:report]{covr::report()}}.
}
