% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ds.R
\name{theme_ds}
\alias{theme_ds}
\title{A theme for ggplot2 with a minimalist design}
\usage{
theme_ds(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{The base font size for text elements in the plot. Default is 11.}

\item{base_family}{The base font family for text elements in the plot. Default is "".}
}
\value{
A ggplot2 theme object with the minimalist design.
}
\description{
The `theme_ds` function is an extension of `theme_bw` from ggplot2
that provides a minimalist design with no major grid lines, minor grid lines,
panel background, and plot background. It is useful for creating simple and clean visualizations.
}
\examples{
library(ggplot2)

# Basic usage
ggplot(mtcars, aes(x = mpg, y = wt)) +
  geom_point() +
  theme_ds()

# Setting font size and font family
ggplot(mtcars, aes(x = mpg, y = wt)) +
  geom_point() +
  theme_ds(base_size = 14, base_family = "Arial")

}
