% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_TIDE_preparing.R
\name{TIDE_prepare}
\alias{TIDE_prepare}
\title{TIDE_prepare}
\usage{
TIDE_prepare(exprset,
keytype="SYMBOL")
}
\arguments{
\item{exprset}{genes(row) * samples(column) data.frame}

\item{keytype}{The type of gene ID. Choose "SYMBOL"(default) or "ENTREZID".}
}
\value{
A data.frame with ENTREZID rownames
}
\description{
A function to prepare the input for TIDEpy
}
