% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_TIMER_immunedeconv.R
\name{TIMER_deconv}
\alias{TIMER_deconv}
\title{TIMER_deconv}
\usage{
TIMER_deconv(exprset,
indication = "PAAD")
}
\arguments{
\item{exprset}{the expression matrix (not log2 tranformed).}

\item{indication}{the tumor type (TCGA) of samples.}
}
\value{
A data.frame containing predicted immune cell proportions.
}
\description{
Deconvolute bulk expression matrix and predict immune cell proportions.
}
\author{
GM.W
}
