% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPI.R
\name{TPI}
\alias{TPI}
\title{Calculates Topographic Position Index}
\usage{
TPI(
  bathy,
  d = c(3, 3),
  na.rm = FALSE,
  use_circle = FALSE,
  unit = "cell",
  pad = FALSE,
  na.rm_type = 1
)
}
\arguments{
\item{bathy}{bathymetry raster}

\item{d}{For rectangular window, it is a vector (or integer) specifying dimensions of window in rows and columns (default is 3x3). For a circular neighborhood this value represents the radius.}

\item{na.rm}{logical. If TRUE, NA will be removed from focal computations. The result will only be NA if all focal cells are NA. (Note: na.rm=TRUE will be much slower). Default is FALSE.}

\item{use_circle}{logical specifying whether or not to use a circular neighborhood (default is FALSE)}

\item{unit}{unit for d if using circular window. Either "cell" (default) or "map"}

\item{pad}{logical. If TRUE, additional 'virtual' rows and columns are padded to x such that there are no edge effects. Only applicable if na.rm=TRUE. Value of virtual cells is set to NA. Default is FALSE}

\item{na.rm_type}{default is 1. If set to 2, when na.rm=FALSE, it will remove all zero weights values from focal window before calculating so that NA's in zero weight cells do not cause calculation to evaluate to NA (Only applicable for circle and na.rm=FALSE, and much slower than na.rm_type=1 with na.rm=FALSE)}
}
\description{
Calculates Topographic Position Index (the difference between a central pixel and the mean of its surrounding cells). Unit is the unit of the input raster. Positive values inidcate a crest, while negative values indicate a depression. 0 value is flat.
}
