% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paircor.R
\name{translit}
\alias{translit}
\title{converts text string in cyrillic into latin text string}
\usage{
translit("моя жизнь", space = "_")
}
\arguments{
\item{str}{is string in cyrillic}
\item{space}{is string to replace space, "_" - by default}
}
\value{
converts text string in cyrillic into latin text string
}
\description{
converts text string in cyrillic into latin text string (letter by letters)
digits and latin symbols stay at place, space is replaced on "space", other symbols are deleted

}

