% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.r
\name{transposedata}
\alias{transposedata}
\title{Transpose data layout}
\usage{
transposedata(df)
}
\arguments{
\item{df}{A dataframe containing abundances of invertebrate taxa in
different samples.}
}
\value{
A data frame transposing the input data, with row and column
labels processed correctly.
}
\description{
Transposes a dataset, correctly processing column and
row labels.
}
\examples{
# transpose the built-in River Almond dataset
# this would have to be transposed back to original format for calculation

transposedata(almond)
}

