% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tread_excel.R
\name{tread_excel}
\alias{tread_excel}
\title{transposing fread}
\usage{
tread_excel(input, sheet = 1, error_as_NA = TRUE,
  rm_blank_col = TRUE)
}
\arguments{
\item{input}{input data}

\item{sheet}{numeric or character to select sheet}

\item{error_as_NA}{turn error as NA (e.g., #DIV/0!)}

\item{rm_blank_col}{boolean to decide removing blank columns}
}
\description{
transposing fread
}
