% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update.R
\name{Update}
\alias{Update}
\title{Update}
\usage{
Update(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of x and y.
}
\description{
This function update the package.
}
\examples{
add_numbers(1, 3)
}
