% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/very_simple_functions.R
\name{utility-functions}
\alias{utility-functions}
\alias{idx_miss}
\alias{n_na}
\alias{is.not.na}
\alias{is.not.null}
\title{Utility Functions for Data Handling}
\usage{
idx_miss(x)

n_na(x)

is.not.na(x)

is.not.null(x)
}
\description{
This documentation provides a common reference for multiple utility functions:
\itemize{
  \item \code{\link{idx_miss}}: Get indices of NA values.
  \item \code{\link{n_na}}: Count the number of NA values.
  \item \code{\link{is.not.na}}: Check for non-NA values.
  \item \code{\link{is.not.null}}: Check for non-NULL values.
}
}
