% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_info.R
\name{video_info}
\alias{video_info}
\title{Gets video details}
\usage{
video_info(
  vid_path,
 
    Media_Info_Path = "C:/Users/socce/Documents/Grad School/Software/Media_Info/MediaInfo.exe"
)
}
\arguments{
\item{vid_path}{file path of video}

\item{Media_Info_Path}{Path to Media Info program}
}
\description{
Uses program Media Info to extract video details
}
