% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wood_aspect.R
\name{wood_aspect}
\alias{wood_aspect}
\title{Helper function for calculating wood aspect}
\usage{
wood_aspect(depth_vector)
}
\arguments{
\item{depth_vector}{passed from raster::focal}
}
\description{
DOES NOT GIVE SIMILAR VALUES TO QGIS. NEED TO CHECK!!!!!!!!!!!!
Helper function for calculating wood aspect
}
