% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wood_slp.R
\name{wood_slp}
\alias{wood_slp}
\title{Helper function for calculating wood slope}
\usage{
wood_slp(depth_vector)
}
\arguments{
\item{depth_vector}{passed from raster::focal}
}
\description{
DOES NOT GIVE SIMILAR VALUES TO QGIS. NEED TO CHECK!!!!!!!!!!!!
Helper function for calculating wood slope
}
