% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{workspace}
\alias{workspace}
\title{Print Elements in Workspace}
\usage{
workspace()
}
\value{
Prints the names of all elements in the global environment to the console.
}
\description{
This function prints the names of all the loaded elements (variables, functions, etc.)
in your global environment.
}
\examples{
# Assume you have variables a and b in your global environment
a <- 1
b <- 2
workspace()  # Output will list 'a' and 'b'
}
