% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRaster2.R
\name{writeRaster2}
\alias{writeRaster2}
\title{Modification of writeRaster that can delete auxillary files}
\usage{
writeRaster2(x, filename, overwrite = TRUE, ...)
}
\arguments{
\item{x}{Raster* object}

\item{filename}{Output filename}

\item{overwrite}{Logical. If TRUE, "filename" will be overwritten if it exists. Additionally, auxillary files related to this raster will be deleted}

\item{...}{other arguments passed to writeRaster}
}
\description{
Modification of writeRaster that can delete auxillary files. If overwrite is set to TRUE (the default) it will overwrite raster file and delete all associated files (e.g. pyramid files). Might only work on Windows.
}
