% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPI.R
\name{BPI}
\alias{BPI}
\title{Calculates Bathymetric Position Index}
\usage{
BPI(bathy, inner_rad, outer_rad, na.rm = FALSE, pad = FALSE,
  na.rm_type = 1, software = "R", Arc_Conn)
}
\arguments{
\item{bathy}{bathymetry raster}

\item{inner_rad}{Inner radius of annulus}

\item{outer_rad}{Outer radius of annulus}

\item{na.rm}{logical. If TRUE, NA will be removed from focal computations. The result will only be NA if all focal cells are NA. (Note: na.rm=TRUE will be much slower). Default is FALSE.}

\item{pad}{logical. If TRUE, additional 'virtual' rows and columns are padded to x such that there are no edge effects. Only applicable if na.rm=TRUE. Value of virtual cells is set to NA. Default is FALSE
#' @param na.rm_type default is 1. If set to 2, when na.rm=FALSE, it will remove all zero weights values from focal window before calculating so that NA's in zero weight cells do not cause calculation to evaluate to NA (Only applicable for na.rm=FALSE, and much slower than na.rm_type=1 with na.rm=FALSE)}

\item{software}{"R" "QGIS" or "Arc" (Note: If using QGIS you must first establish a connection using RQGIS3::set_env() and RQGIS3::qgis_session_info(). If using Arc you must have a Spatial Analyst license. Also for either the R Version architecture type (32bit or 64bit) must match architecture type of ArcGIS/QGIS)}

\item{Arc_Conn}{Connection to ARcGIS. object output by RPyGeo::rpygeo_build_env. Make sure to set extensions argument to "Spatial"}
}
\description{
Calculates Bathymetric Position Index (the difference between a central pixel and the mean of the values in the specified annulus). Positive values inidcate a crest, while negative values indicate a depression. 0 value is flat.
}
