% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPI.R
\name{BPI}
\alias{BPI}
\title{Calculates Bathymetric Position Index}
\usage{
BPI(bathy, inner_rad, outer_rad, rad_units = "cell",
  round_values = FALSE, standardize = FALSE, na.rm = FALSE,
  parallel = FALSE, ncores = parallel::detectCores() - 1)
}
\arguments{
\item{bathy}{bathymetry raster}

\item{inner_rad}{Inner radius of annulus}

\item{outer_rad}{Outer radius of annulus}

\item{round_values}{logical specifying whether or not to round values to nearest integer (FALSE by default, but BTM algorithm in ArcGIS rounds by default)}

\item{standardize}{logical specifying whether or not to "standardize the BPI grid" (z-score). Default is FALSE}

\item{parallel}{logical specifying whether or not to run in parallel}

\item{ncores}{number of cores to use if running in parallel}
}
\description{
Calculates Bathymetric Position Index (the difference between a central pixel and the mean of the values in the specified annulus). Positive values inidcate a crest, while negative values indicate a depression. 0 value is flat.
CURRENTLY DOES NOT MATCH VALUES FROM BENTHIC TERRAIN MODELLER
}
