% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_width.R
\name{calc_width}
\alias{calc_width}
\title{Calculates width of video frame}
\usage{
calc_width(alt, pitch, cam_angle = 32.8, HFOV_air = 82.4)
}
\arguments{
\item{alt}{altitude}

\item{pitch}{pitch}

\item{cam_angle}{camera angle in degrees}

\item{HFOV_air}{HFOV in air for given camera}
}
\description{
Calculates width of video frame in meters
}
