% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_reset.R
\name{check_if_reset}
\alias{check_if_reset}
\title{Checks if camera has been restarted during transect}
\usage{
check_if_reset(cam_filepath, display_warning = TRUE,
  return_idx = FALSE)
}
\arguments{
\item{cam_filepath}{vector of filepaths from blackfly camera}

\item{display_warning}{Logical indicating whether a warning should be displayed if restart is detected}

\item{return_idx}{Logical indicating whether the index of where the restart occurred should also be returned}
}
\description{
Checks if camera has been restarted during transect by seeing if frames where written to different directories. Optionally can return index where restart occurred (Gives index of first value after restart).
}
