% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_reset.R
\name{check_if_reset}
\alias{check_if_reset}
\title{Checks if blackfly has been restarted during transect}
\usage{
check_if_reset(cam_filepath)
}
\arguments{
\item{cam_filepath}{vector of filepaths from blackfly camera}
}
\description{
Does that by checking if gigegrab number from file decreases at any point
}
