% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blackfly_timestamp_from_vidtime.R
\name{check_if_reset}
\alias{check_if_reset}
\title{Checks if camera has been restarted during transect}
\usage{
check_if_reset(cam_filepath)
}
\arguments{
\item{cam_filepath}{vector of filepaths from balckfly camera (must be same length as cam_timestamp)}
}
\description{
Does that by checking if gigegrab number from file decreases at any point
}
