% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_theme.R
\name{custom_theme}
\alias{custom_theme}
\title{Title}
\usage{
custom_theme(base_size = 9)
}
\arguments{
\item{base_size}{Base text size}
}
\value{
a theme plot
}
\description{
This is my theme and it is the Best
}
