% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_cam_pattern.R
\name{detect_cam_pattern}
\alias{detect_cam_pattern}
\title{Function for detecting string type in camera table}
\usage{
detect_cam_pattern(cam_filepath)
}
\arguments{
\item{cam_filepath}{vector of filepaths from blackfly camera}
}
\description{
Function for detecting string type in camera table
}
