% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_info.r
\name{environment_info}
\alias{environment_info}
\title{Environment Info}
\usage{
environment_info(msg)
}
\arguments{
\item{msg}{The message that should be printed}
}
\description{
A function to print information about the current environment.
}
\details{
This function prints current environment information, and a message.
}
\examples{
environment_info("This is an important message from your sponsor.")
}
\keyword{debugging}
