% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_info.R
\name{environment_info}
\alias{environment_info}
\title{this prints the environment}
\usage{
environment_info(msg)
}
\arguments{
\item{msg}{A debugging message to be printed}
}
\description{
this prints the environment
}
