% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{environment_info}
\alias{environment_info}
\title{A function to print information about the current environment.}
\usage{
environment_info(msg)
}
\arguments{
\item{msg}{The message that should be printed}
}
\description{
This function prints current environment information, and a message.
}
\examples{
environment_info("Hi, what is your name?")
}
\keyword{debugging}

