% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_info.R
\name{environment_info}
\alias{environment_info}
\title{Environment info}
\usage{
environment_info(msg)
}
\arguments{
\item{msg}{the message to be printed}
}
\description{
This prints the session and a message
}
\examples{
environment_info("Hello!")
}
