% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment_info.R
\name{environment_info}
\alias{environment_info}
\title{Environment Info}
\usage{
environment_info(msg)
}
\arguments{
\item{msg}{The message to be pringted}
}
\description{
This prints the session and a message
}
\examples{
enviornment_info("hello!")
}
