% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blackfly_timestamp_from_vidtime.R
\name{get_blackfly_timestamp_from_vidtime}
\alias{get_blackfly_timestamp_from_vidtime}
\title{Returns Blackfly Timestamp based on Video Number and Seconds in}
\usage{
get_blackfly_timestamp_from_vidtime(video_col, seconds_col,
  Total_Frame = NULL, cam_timestamp, cam_filepath, test = FALSE,
  vid_length = 1, fps = calc_fps(cam_timestamp, cam_filepath),
  round_fps = FALSE)
}
\arguments{
\item{video_col}{vector of video numbers (must be same length as seconds_col)}

\item{seconds_col}{vector of seconds in video (must be same length as video_col)}

\item{Total_Frame}{Total Frame Number can be used instead of video_col and seconds_col}

\item{cam_timestamp}{vector of camera timestamps (must be same length as cam_filepath)}

\item{cam_filepath}{vector of filepaths from balckfly camera (must be same length as cam_timestamp)}

\item{test}{if TRUE output will be a tibble that can be used to check if the function is working correctly}

\item{vid_length}{Length of each video in minutes}

\item{fps}{Frames Per Second (Can be calculated by default from cam_timestamp and cam_filepath)}

\item{round_fps}{Logical value specifying whether or not to round frames per second to nearest integer. This is FALSE by default. Some Blackfly videos (mostly older ones) were sequenced using an approximate integer value fps causing it to prgressively drift from axis cameras. If the resulting timestamps seem to provide a value off by a few seconds (especially towards the ends of transects), try switching this value.}
}
\description{
Returns a vector of Blackfly Timestamps corresponding to video_col and video_sec by using those two parameters to approximate frame number (or can use Total Frame Number directly) and then matches that frame number to a table to see when that frame was recorded.
}
