% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blackfly_timestamp_from_vidtime.R
\name{get_blackfly_timestamp_from_vidtime}
\alias{get_blackfly_timestamp_from_vidtime}
\title{Returns Blackfly Timestamp based on Video Number and Seconds in}
\usage{
get_blackfly_timestamp_from_vidtime(video_col, seconds_col, cam_timestamp,
  cam_filepath)
}
\arguments{
\item{video_col}{vector of video numbers (must be same length as seconds_col)}

\item{seconds_col}{vector of seconds in video (must be same length as video_col)}

\item{cam_timestamp}{vector of camera timestamps (must be same length as cam_filepath)}

\item{cam_filepath}{vector of filepaths from blackfly camera (must be same length as cam_timestamp)}
}
\description{
Returns a vector of Blackfly Timestamps corresponding to video_col and video_sec by using those two parameters to approximate frame number and then match that frame number to a table to see when that frame was recorded. Timestamps and filepaths vectors must correspond to eachother and be in the order that the frames were recorded
}
