% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_payout.R
\name{get_payout}
\alias{get_payout}
\title{Gets Payout data in single table from winch data folder}
\usage{
get_payout(winch_dir, trim_rows = TRUE)
}
\arguments{
\item{winch_dir}{directory containing winch data}

\item{trim_rows}{A logical that if TRUE (default), will take the median of all values with the same timestamp so that each timestamp has one value}
}
\description{
Gets Payout data in single table from winch data folder.
}
