% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSNPseq.R
\name{getSNPseq}
\alias{getSNPseq}
\title{getSNPseq}
\usage{
getSNPseq(snp.vcf, genome, n_flank, IUPAC = FALSE)
}
\arguments{
\item{snp.vcf}{A table contain the SNP information. The format is the first five columns of the VCF format plus a column for strand. (CHROM,POS,ID,REF,ALT,strand).}

\item{genome}{The BSgenome object}

\item{n_flank}{The number of nucleotide to get around the SNP.}

\item{IUPAC}{logcal option. Whether to output IUPAC code for SNP instead of outputing two sequence.}
}
\description{
Given BSgenome object, get the sequence flanking the SNPs.
}
