% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp.R
\name{interp}
\alias{interp}
\title{Interpolate Ship Position}
\usage{
interp(my_data, na.rm = FALSE)
}
\arguments{
\item{my_data}{vector of data}

\item{na.rm}{Logical indicating whether a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\description{
Interpolates NA's by filling them with the average of the previous and subsequent values.
}
