% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_cvision_fish2hab.R
\name{match_cvision_fish2hab}
\alias{match_cvision_fish2hab}
\title{Matches fish recorded in C-Vision to closest habitat frame}
\usage{
match_cvision_fish2hab(hab_totalframe, merged_cvision_csv,
  include_framenum = FALSE)
}
\arguments{
\item{hab_totalframe}{vector of total frame number of habitat observation}

\item{merged_cvision_csv}{output of merge_cvision_csv}

\item{include_framenum}{whether or not to include frame number in output table}
}
\description{
Matches fish recorded in C-Vision to closest habitat frame. Returns a tibble of species by site counts.
}
