% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytools-package.R
\docType{package}
\name{mytools-package}
\alias{mytools}
\alias{mytools-package}
\title{mytools: Set of Useful R Functions}
\description{
This package provides utility functions for R. These functions include creating graphs and summary statistics.
}
\author{
\strong{Maintainer}: Caleb Henning \email{chenning@wesleyan.edu}

}
\keyword{internal}
