% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytools-package.R
\docType{package}
\name{mytools-package}
\alias{mytools}
\alias{mytools-package}
\title{mytools: A set of R utility function for me}
\description{
Fast statistics and graphs using the tidyverse.
}
\author{
\strong{Maintainer}: Larissa Xu \email{lxu01@wesleyan.edu}

}
\keyword{internal}
