% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytools-package.R
\docType{package}
\name{mytools-package}
\alias{mytools}
\alias{mytools-package}
\title{mytools: A set of R utility functions for me}
\description{
Simple package, statistics and a graph
}
\author{
\strong{Maintainer}: Tegh Singh \email{tsingh@wesleyan.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
