% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_bins.R
\name{order_bins}
\alias{order_bins}
\title{Order Percent Cover Bins}
\usage{
order_bins(bins)
}
\arguments{
\item{bins}{vector of bin ranges}
}
\description{
Orders Percent Cover Bins in ascending order based on lower range
}
