% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perprim2perhard.R
\name{perprim2perhard}
\alias{perprim2perhard}
\title{Percent Primary to Percent Hardbottom}
\usage{
perprim2perhard(indur_col, per_col)
}
\arguments{
\item{indur_col}{vector of induration}

\item{per_col}{vector of percent primary cover bins}
}
\description{
Takes a vector of Percent Primary Cover and outputs a vector of Percent Hardbottom
}
\keyword{percent_cover}
