% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qscatter.R
\name{qscatter}
\alias{qscatter}
\title{Quick Scatter Plot}
\usage{
qscatter(data, x, y)
}
\arguments{
\item{data}{a data frame}

\item{x}{a numeric variable}

\item{y}{a numeric variable}
}
\value{
a ggplot 2 graph
}
\description{
Scatter plot with lines of best fit, correlation coefficient, and p-value
}
\examples{
qscatter(mtcars, wt, mpg)
}
