% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qscatter.R
\name{qscatter}
\alias{qscatter}
\title{Quick scatter plot}
\usage{
qscatter(data, x, y)
}
\arguments{
\item{data}{a data frame}

\item{x}{a numeric variable}

\item{y}{a numeric variable}
}
\value{
a tibble with n, mean, and standard deviation
}
\description{
Create scatter plot with linear line of best fit, correlation, and p-value
}
\examples{
qscatter(mtcars, wt, hp)

}
