% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qscatter.R
\name{qscatter}
\alias{qscatter}
\title{Quick Scatter Plot}
\usage{
qscatter(data, x, y)
}
\arguments{
\item{data}{a data frame}

\item{x}{a numeric variable}

\item{y}{a numeric variable}
}
\value{
a ggplot2 graph
}
\description{
Scatter plot with line of best fit and correlation
}
\examples{
qscatter(mtcars, wt, hp)
}
