% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_save_ggplot.R
\name{save_ggplot}
\alias{save_ggplot}
\title{Save the Last ggplot Object Quietly}
\usage{
save_ggplot(filename, subdomain = "plots")
}
\arguments{
\item{filename}{The name of the file to save the plot as. Do not include the file extension.}

\item{subdomain}{The subdirectory within the current working directory where the plot will be saved. Default is 'plots'.}
}
\value{
NULL. The function is called for its side effect of saving a plot.
}
\description{
This function saves the last ggplot object to a specified subdirectory
within the current working directory. The function suppresses any messages
generated during the save operation.
}
\examples{
\dontrun{
# Create a simple ggplot object
p <- ggplot(mtcars, aes(x = mpg, y = wt)) + geom_point()
# Display the plot
print(p)
# Save the plot
save_ggplot("my_plot")
}
}
