% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_tools.r
\name{summarySD}
\alias{summarySD}
\title{Summary of data}
\usage{
summarySD(data = NULL, measurevar, groupvars, error = "sd", na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE)
}
\arguments{
\item{data}{A data frame containing the data to be summarised.}

\item{measurevar}{The name of a column that contains the variable to be 
summarized.}

\item{groupvars}{A vector containing names of columns that contain 
grouping variables.}

\item{error}{Error type (sd, se or ci) to be used. Defaults to sd.}

\item{na.rm}{A boolean that indicates whether to ignore NA's}

\item{conf.interval}{The percent range of the confidence interval (default 
is 95%).}
}
\value{
A summary of the data, with the required measure of variability.
}
\description{
Gives count, mean, standard deviation, standard error of the 
mean, and confidence interval (default 95%).
}

