% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPI.R
\name{TPI}
\alias{TPI}
\title{Calculates Topographic Position Index}
\usage{
TPI(bathy, window_size = c(3, 3), na.rm = FALSE, use_circle = FALSE)
}
\arguments{
\item{bathy}{bathymetry raster}

\item{window_size}{For rectangular window, it is a vector (or integer) specifying dimensions of window in rows and columns (default is 3x3). Window size must be odd. For a circular neighborhood this value represents diameter in map units (e.g. meters).}

\item{use_circle}{logical specifying whether or not to use a circular neighborhood (default is FALSE)}
}
\description{
Calculates Topographic Position Index (the difference between a central pixel and the mean of its surrounding cells). Unit is the unit of the input raster. Positive values inidcate a crest, while negative values indicate a depression. 0 value is flat.
}
